% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-dataPreparation.R
\name{addAndValidateErrorModel}
\alias{addAndValidateErrorModel}
\title{Add and Validate Error Model}
\usage{
addAndValidateErrorModel(
  projectConfiguration = projectConfiguration,
  dtPrior,
  dataObservedForMatch
)
}
\arguments{
\item{projectConfiguration}{An object of class ProjectConfiguration containing project configuration details.}

\item{dtPrior}{A data.table containing prior definitions.}

\item{dataObservedForMatch}{A data.table containing observed data for matching.}
}
\value{
The data.table \code{dataObservedForMatch} with merged error model information.
}
\description{
This function adds and validates the error model based on the project configuration and observed data.
}
\keyword{internal}
