% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-template.R
\name{addBMLMWorkflowTemplate}
\alias{addBMLMWorkflowTemplate}
\title{Opens the workflow template as new document}
\usage{
addBMLMWorkflowTemplate()
}
\description{
Opens the workflow template as new document
}
\seealso{
Other configuration: 
\code{\link{addBMLMPConfiguration}()},
\code{\link{configurePriors}()},
\code{\link{convertVersionBMLM}()},
\code{\link{getCurrentConfigTable}()},
\code{\link{loadListsForRun}()},
\code{\link{readIdentificationParameterFromSnapshot}()}
}
\concept{configuration}
