% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-config.R
\name{addGlobalPriorParameter}
\alias{addGlobalPriorParameter}
\title{Add Global Prior Parameter}
\usage{
addGlobalPriorParameter(dtPrior, dtDefinition)
}
\arguments{
\item{dtPrior}{A data.table containing prior parameters.}

\item{dtDefinition}{A data.table containing parameter definitions.}
}
\value{
A data.table containing updated prior parameters.
}
\description{
This function adds global prior parameters based on the parameter definition.
}
\keyword{internal}
