% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-BMLMParameter.R
\name{addMinMaxValues}
\alias{addMinMaxValues}
\title{Add Min and Max Values}
\usage{
addMinMaxValues(dt, dtDefinition)
}
\arguments{
\item{dt}{A data.table to which MinValue and MaxValue will be added.}

\item{dtDefinition}{A data.table containing parameter definitions.}
}
\value{
A data.table with MinValue and MaxValue added.
}
\description{
This function merges a data.table with min and max values from a definition data.table.
}
