% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-config.R
\name{addModelErrorParameter}
\alias{addModelErrorParameter}
\title{Add Model Error Parameter}
\usage{
addModelErrorParameter(dtPrior, wb)
}
\arguments{
\item{dtPrior}{A data.table containing prior parameters.}

\item{wb}{the open workkbook with the BMLM configurations}
}
\value{
A data.table containing updated prior parameters.
}
\description{
This function adds model error parameters based on output mappings.
}
\keyword{internal}
