% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotParameter.R
\name{adjustLimitsPlot}
\alias{adjustLimitsPlot}
\title{Adjust Limits for a ggplot Object}
\usage{
adjustLimitsPlot(plotObject, colorScalingVector, titeltxt)
}
\arguments{
\item{plotObject}{A ggplot object that will be modified.}

\item{colorScalingVector}{A named vector of colors for different statuses.
This vector will be used to customize the color scale of the plot.}

\item{titeltxt}{A string to include in the plot title.}
}
\value{
A modified ggplot object with adjusted limits, colors, and labels.
}
\description{
This function modifies a ggplot object by adding horizontal lines at y=0 and y=1,
customizing the color scale, and adjusting the labels and theme.
}
\keyword{internal}
