% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-startBMLMOptimization.R
\name{analyseInitalSimulationFailures}
\alias{analyseInitalSimulationFailures}
\title{Analyze Initial Simulation Failures}
\usage{
analyseInitalSimulationFailures(dtList, optimEnv, loglikelihoods, outputDir)
}
\arguments{
\item{dtList}{A list containing data tables relevant to the prior model and results.}

\item{optimEnv}{An environment object that contains optimization results.}

\item{loglikelihoods}{A named vector containing log-likelihood values}

\item{outputDir}{A character string specifying the directory where logs and outputs
should be saved.}
}
\value{
Returns nothing (invisible NULL). The function performs logging and updates
the model error in the provided data tables.
}
\description{
This function analyzes the results of initial simulations to identify any failures
based on the log-likelihood values. It checks for individuals with missing predicted
values and infinite log-likelihoods, updating the model error as necessary.
}
\keyword{internal}
