% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCorrelation.R
\name{analyzeLabelCorrelations}
\alias{analyzeLabelCorrelations}
\title{Analyze Label Correlations}
\usage{
analyzeLabelCorrelations(mergedData, label, corCut, method, columnVector)
}
\arguments{
\item{mergedData}{A data.table containing merged plot and population data.}

\item{label}{A character string representing the label to analyze.}

\item{corCut}{A numeric value for the correlation cutoff threshold.}

\item{method}{A character string specifying the correlation method to use.}

\item{columnVector}{A caharacter vector with column Names to analyse.}
}
\value{
A list of correlation plots for the specified label.
}
\description{
This helper function analyzes correlations for a given label against all population columns.
}
