% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCorrelation.R
\name{analyzeLabelKruskal}
\alias{analyzeLabelKruskal}
\title{Analyze Label Kruskal-Wallis Tests}
\usage{
analyzeLabelKruskal(mergedData, label, pValueCut, columnVector)
}
\arguments{
\item{mergedData}{A data.table containing merged plot and population data.}

\item{label}{A character string representing the label to analyze.}

\item{pValueCut}{A numeric value for the Chi-squared cutoff threshold.}

\item{columnVector}{A caharacter vector with column Names to analyse.}
}
\value{
A list of Kruskal-Wallis test plots for the specified label.
}
\description{
This helper function performs Kruskal-Wallis tests for a given label against all population columns.
}
