% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-enum.R
\docType{data}
\name{BMLMSHEET}
\alias{BMLMSHEET}
\title{Sheet names of BMLM COnfiguration xlsx}
\format{
An object of class \code{list} of length 7.
}
\usage{
BMLMSHEET
}
\description{
Sheet names of BMLM COnfiguration xlsx
}
\seealso{
Other enum helpers: 
\code{\link{ERRORMODEL}},
\code{\link{PARAMETERTYPE}}
}
\concept{enum helpers}
\keyword{datasets}
