% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotConvergence.R
\name{calculateConvergenceMetrics}
\alias{calculateConvergenceMetrics}
\title{Calculate Convergence Metrics}
\usage{
calculateConvergenceMetrics(dt)
}
\arguments{
\item{dt}{A data.table containing convergence data with necessary columns.}
}
\value{
A data.table with additional columns for objective values and percentages of failure and outside range.
}
\description{
This function calculates the objective values and percentages for the convergence data.
}
\keyword{internal}
