% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-loglikelihood.R
\name{calculateResidual}
\alias{calculateResidual}
\title{Calculate Residuals from Observed and Predicted Values}
\usage{
calculateResidual(yValue, predicted, model, sigma, isCensored, lloq)
}
\arguments{
\item{yValue}{A numeric value representing the observed value.}

\item{predicted}{A numeric value representing the predicted value.}

\item{model}{A character string specifying the model to use for calculating the residuals
(options are "absolute", "proportional", "log_absolute").}

\item{sigma}{A numeric value representing the standard deviation of the residuals.}

\item{isCensored}{A logical value indicating whether the data is censored.}

\item{lloq}{A numeric value representing the lower limit of quantification.}
}
\value{
A numeric value representing the calculated residual.
}
\description{
This function calculates the residuals based on the observed and predicted values,
applying different models for the calculation depending on whether the data is censored.
}
\keyword{internal}
