% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-distribution.R
\name{calculateValueOfDistributionRow}
\alias{calculateValueOfDistributionRow}
\title{Calculate Probability Based on Distribution}
\usage{
calculateValueOfDistributionRow(row, type, value, log)
}
\arguments{
\item{row}{A named vector representing a row from a data frame. It must contain:
\itemize{
\item \code{Distribution}: A string that specifies the distribution type (e.g., "norm" for normal distribution).
\item \code{StartValue}: A numeric value representing the starting point for the probability calculation.
\item Additional parameters with names ending in \verb{_type} and \verb{_value} that specify the distribution parameters.
}}
}
\value{
A numeric value representing the calculated probability. Returns NA in case of an error.
}
\description{
This function calculates the probability for a given distribution using specified parameters.
It extracts the distribution type and its parameters from a data frame row, constructs the
appropriate function call, and returns the calculated probability. If an error occurs during
calculation, it returns NA and prints an error message.
}
