% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertVersion.R
\name{checkAndUpdateScalingMethod}
\alias{checkAndUpdateScalingMethod}
\title{Check and Update Scaling Method}
\usage{
checkAndUpdateScalingMethod(folder)
}
\arguments{
\item{folder}{A character string specifying the path to the folder containing
the status files.}
}
\value{
None
}
\description{
Version 0.1.9 introdcuded a \code{scalingMethod}, which was saved in the status files.
This function checks for the presence of the 'scalingMethod' in status files
within a specified folder. If it is not present, it updates the status file
by adding the \code{scalingMethod} available for older versions
and initializing the 'outsideRangeCounter'.
}
