% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-startBMLMOptimization.R
\name{checkBMLMConfiguration}
\alias{checkBMLMConfiguration}
\title{Check BMLM Configuration}
\usage{
checkBMLMConfiguration(projectConfiguration)
}
\arguments{
\item{projectConfiguration}{A list containing project configuration details.}
}
\value{
NULL if the configuration is valid; stops execution if invalid.
}
\description{
Validates the presence of the BMLM configuration file in the project configuration.
}
\keyword{internal}
