% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-checkProgress.R
\name{checkConvergence}
\alias{checkConvergence}
\title{Check Convergence of Model Parameters}
\usage{
checkConvergence(
  projectConfiguration,
  runName,
  nPoints = 200,
  selectionMode = c("last", "random", "first")
)
}
\arguments{
\item{projectConfiguration}{A list containing configuration settings for the project,
including the output folder path.}

\item{runName}{A character string specifying the name of the run.}

\item{nPoints}{An integer specifying the number of points to select for plotting.
Default is 200.}

\item{selectionMode}{A character string indicating the mode of selection for points.
Options are 'last', 'random', and 'first'. Default is 'last'.}
}
\value{
A ggplot object visualizing the convergence of model parameters.
The plot shows the values of logTimeProfile, logHyperParameter, logPrior,
and objectiveValue over iterations.
}
\description{
This function checks the convergence of model parameters by reading a CSV file
containing convergence data and visualizing the results using ggplot2.
}
\examples{
\dontrun{
# Example usage:
plot <- checkConvergence(projectConfiguration, "run1")
print(plot)
}

}
