% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-dataPreparation.R
\name{checkDuplicates}
\alias{checkDuplicates}
\title{Check for Duplicates in Data Frames}
\usage{
checkDuplicates(dt, identifierCols, sheetName)
}
\arguments{
\item{dt}{A data.table to check for duplicates.}

\item{identifierCols}{A character vector of column names to check for uniqueness.}

\item{sheetName}{A character string representing the name of the sheet for error messages.}
}
\value{
NULL. The function stops execution if duplicates are found.
}
\description{
This function checks for duplicates in specified columns of a data frame
and throws an error if duplicates are found. It generates a descriptive
error message indicating which parameters are duplicated.
}
\keyword{internal}
