% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-checkProgress.R
\name{checkParameterValues}
\alias{checkParameterValues}
\title{Check Parameter Values}
\usage{
checkParameterValues(projectConfiguration, runName)
}
\arguments{
\item{projectConfiguration}{A list containing configuration settings for the project,
including the output folder path.}

\item{runName}{A character string specifying the name of the run.}
}
\value{
A list of ggplot objects visualizing the parameter values.
Each plot shows the current, best, and starting values of parameters.
}
\description{
This function checks the values of model parameters by reading status files
and visualizing the current, best, and starting values using ggplot2.
}
\examples{
\dontrun{
# Example usage:
plots <- checkParameterValues(projectConfiguration, "run1")
lapply(plots, print)
}

}
