% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-distribution.R
\name{computeStatFunction}
\alias{computeStatFunction}
\title{Compute Statistical Function Values}
\usage{
computeStatFunction(
  values,
  parameters,
  distribution,
  v,
  type = c("D", "P", "Q", "R"),
  log = FALSE,
  normalisationFactor = 1
)
}
\arguments{
\item{values}{A numeric vector of parameter values for the specified distribution.}

\item{parameters}{A character vector of parameter names corresponding to the values.}

\item{distribution}{A character string specifying the distribution (e.g., "norm", "lnorm").}

\item{v}{A numeric value or vector for which the function should compute the result,
depending on the specified type.}

\item{type}{A character string indicating the type of statistical function to compute.
It can be one of "P" (probability), "D" (density), "Q" (quantile), or "R" (random generation).}

\item{log}{A logical value indicating whether to return the logarithm of the probability
or density (if applicable). Default is FALSE.}

\item{normalisationFactor}{A numeric value used to adjust the computed densities for
truncated distributions. Default is 1 (no adjustment).}
}
\value{
The result of the statistical function call, which varies based on the type:
\describe{
\item{P}{Probability value for the given quantile.}
\item{D}{Density value for the given input.}
\item{Q}{Quantile value for the given probability.}
\item{R}{Random samples generated from the specified distribution.}
}
}
\description{
This function evaluates statistical functions (probability, density, quantile,
and random generation) based on the specified distribution and parameters.
}
\examples{
# Example usage:
computeStatFunction(values = c(0, 1), parameters = c("mean", "sd"),
                     distribution = "norm", v = 0.5, type = "P")

}
