% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertVersion.R
\name{convertVersionBMLM}
\alias{convertVersionBMLM}
\title{Roxygenheader}
\usage{
convertVersionBMLM(projectConfiguration)
}
\arguments{
\item{projectConfiguration}{A list containing configuration settings for the project.}
}
\value{
None
}
\description{
Update files generated by ospsuite.bmlm
}
\details{
This script converts files generated with old package versions
so that they can run with newer versions by updating the scaling method.
}
