% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-config.R
\name{createBMLMProjectConfiguration}
\alias{createBMLMProjectConfiguration}
\title{Title}
\usage{
createBMLMProjectConfiguration(
  projectConfigurationFilePath,
  bMLMConfigurationFile
)
}
\arguments{
\item{overwrite}{}
}
\description{
Title
}
