% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-template.R
\name{createBMLMWorkflowTemplate}
\alias{createBMLMWorkflowTemplate}
\title{Opens the workflow template as new document}
\usage{
createBMLMWorkflowTemplate()
}
\description{
Opens the workflow template as new document
}
