% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCorrelation.R
\name{createCorrelationPlot}
\alias{createCorrelationPlot}
\title{Create Correlation Plot}
\usage{
createCorrelationPlot(data, label, popCol, correlationValue)
}
\arguments{
\item{data}{A data.table containing the data for plotting.}

\item{label}{A character string representing the label on the y-axis.}

\item{popCol}{A character string representing the population column on the x-axis.}

\item{correlationValue}{A numeric value representing the correlation coefficient.}
}
\value{
A ggplot object representing the correlation plot.
}
\description{
This helper function generates a scatter plot with a linear regression line for the correlation between two variables.
}
