% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotParameter.R
\name{createDistributionTables}
\alias{createDistributionTables}
\title{Create Distribution Tables}
\usage{
createDistributionTables(dtPrior, hyperParameter, plotList, dtValues)
}
\arguments{
\item{dtPrior}{A data.table containing prior information.}

\item{hyperParameter}{A data.table containing hyperparameter information.}

\item{plotList}{A list of plots to be generated.}

\item{dtValues}{A data.table containing values for plotting.}
}
\value{
A list of distribution tables.
}
\description{
This function generates tables for hyperparameters and their distributions.
}
\keyword{internal}
