% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-dataPreparation.R
\name{createDtList}
\alias{createDtList}
\title{Create Data List for BMLM Optimization}
\usage{
createDtList(projectConfiguration, scenarioList, dataObserved, seed)
}
\arguments{
\item{projectConfiguration}{An object of class ProjectConfiguration containing project configuration details.}

\item{scenarioList}{A list of scenarios to be used in the optimization.}

\item{dataObserved}{A data.table containing observed data.}

\item{seed}{A numeric seed for random number generation.}
}
\value{
A list of data.tables required for the BMLM optimization process.
}
\description{
This helper function creates a list of data.tables (\code{dtList}) required for the BMLM optimization process.
}
\keyword{internal}
