% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCorrelation.R
\name{createKruskalPlot}
\alias{createKruskalPlot}
\title{Create Kruskal-Wallis Plot}
\usage{
createKruskalPlot(data, label, popCol, pValue)
}
\arguments{
\item{data}{A data.table containing the data for plotting.}

\item{label}{A character string representing the label on the y-axis.}

\item{popCol}{A character string representing the population column on the x-axis.}

\item{pValue}{A numeric value representing the p-value from the Kruskal-Wallis test.}
}
\value{
A ggplot object representing the Kruskal-Wallis plot.
}
\description{
This helper function generates a box-and-whisker plot for the Kruskal-Wallis test results.
}
