% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-startBMLMOptimization.R
\name{createObjectiveFunction}
\alias{createObjectiveFunction}
\title{Create Objective Function for Optimization}
\usage{
createObjectiveFunction(
  optimEnv,
  dtList,
  scenarioList,
  simulationRunOptions,
  outputDir,
  lastStatusSavingIntervalInSecs,
  withInternalOptimization
)
}
\arguments{
\item{optimEnv}{An environment containing the current state variables for optimization.}

\item{dtList}{A list containing data tables relevant for the optimization process.}

\item{scenarioList}{A list of scenarios to be optimized.}

\item{simulationRunOptions}{An obkect of class SimulationRunOptions.}

\item{outputDir}{A character string representing the path to the output directory.}

\item{lastStatusSavingIntervalInSecs}{An integer specifying the interval for saving the last status in seconds.}
}
\value{
A function that takes a vector of parameters and returns the current evaluation of the objective function.
}
\description{
Generates an objective function to be used in the optimization process.
}
\details{
This function returns a closure that encapsulates the optimization logic,
allowing the \code{optim} function to evaluate the objective based on the given
parameters and the current optimization environment.

The objective function updates the optimization environment, evaluates time profiles,
performs internal optimization, and updates the optimization status. It also handles errors
during the optimization process.
}
\keyword{internal}
