% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotParameter.R
\name{createPriorLineData}
\alias{createPriorLineData}
\title{Create Prior Line Data}
\usage{
createPriorLineData(dtPrior)
}
\arguments{
\item{dtPrior}{A data.table containing prior information.}
}
\value{
A data.table containing the line data for prior distributions.
}
\description{
This function generates line data for prior distributions.
}
\keyword{internal}
