% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-config.R
\name{createStartValues}
\alias{createStartValues}
\title{Create Start Values}
\usage{
createStartValues(dtDefinition, dataObserved)
}
\arguments{
\item{dtDefinition}{A data.table containing parameter definitions, including
names and units for individual parameters.}

\item{dataObserved}{A data.table of observed data.}
}
\value{
A data.table containing start values for each individual parameter.
}
\description{
This function generates start values for individual parameters based on the
parameter definitions and unique individual IDs from the observed data.
}
\keyword{internal}
