% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotParameter.R
\name{defaultColorsParameterPlots}
\alias{defaultColorsParameterPlots}
\title{Get Default Colors for Parameter Plots}
\usage{
defaultColorsParameterPlots()
}
\value{
A named vector of colors for different statuses. The default colors
are:
\itemize{
\item current: "darkgreen"
\item start: "lightblue"
\item best: "orange"
}
}
\description{
This function retrieves the default color settings for parameter plots.
It checks the global option \code{ospsuite.BMLM.defaultColors} and returns the
specified colors. If no option is set, it returns a predefined set of colors.
}
\examples{
\dontrun{
# Set custom colors for parameter plots
options(ospsuite.BMLM.defaultColors = c(current = "blue", start = "yellow", best = "red"))
customColors <- defaultColorsParameterPlots()
print(customColors)
}
}
