% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-startBMLMOptimization.R
\name{evaluateInitialValues}
\alias{evaluateInitialValues}
\title{Evaluate Model at Initial Values}
\usage{
evaluateInitialValues(dtList, outputDir, scenarioList)
}
\arguments{
\item{dtList}{A data table containing the necessary data for the evaluation, including prior and start values.}

\item{outputDir}{The directory where the log file will be saved.}
}
\value{
NULL This function does not return any value but logs information to the console and a log file.
}
\description{
This function evaluates initial values for a given project configuration by running simulations for specified scenarios.
It logs the progress and results to both the console and a designated log file.
}
\seealso{
Other optimization: 
\code{\link{calculateLogLikelihood}()},
\code{\link{optimizeParameters}()},
\code{\link{processScenario}()}
}
\concept{optimization}
