% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-startBMLMOptimization.R
\name{evaluateLogLikelihood}
\alias{evaluateLogLikelihood}
\title{Evaluate Log Likelihood}
\usage{
evaluateLogLikelihood(loglikelihoods, optimEnv, outputDir)
}
\arguments{
\item{loglikelihoods}{A vector of log likelihood values.}

\item{optimEnv}{The optimization environment containing current state variables.}

\item{outputDir}{A character string representing the path to the output directory.}
}
\value{
A numeric value representing the current evaluation of the objective function.
}
\description{
Evaluates the log likelihoods and returns the current value based on the
provided log likelihoods.
}
\keyword{internal}
