% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-startBMLMOptimization.R
\name{evaluateTimeprofiles}
\alias{evaluateTimeprofiles}
\title{Evaluate Time Profiles}
\usage{
evaluateTimeprofiles(
  optimEnv,
  scenarioList,
  dtList,
  simulationRunOptions,
  withProtocol = FALSE,
  outputDir = NULL
)
}
\arguments{
\item{optimEnv}{The optimization environment containing current state variables.}

\item{scenarioList}{A list of scenarios to be evaluated.}

\item{simulationRunOptions}{An obkect of class SimulationRunOptions.}

\item{dtPrior}{A data.table containing prior values.}

\item{dtStartValues}{A data.table containing start values.}

\item{dtMappedPaths}{A data.table containing mapped paths for parameters.}

\item{dataObservedForMatch}{A data.table containing observed data for matching.}
}
\value{
NULL (invisible).
}
\description{
Updates parameter values and runs results for each scenario.
}
\keyword{internal}
