% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCorrelation.R
\name{excludeIdenticalFactors}
\alias{excludeIdenticalFactors}
\title{Exclude Identical Factor Columns}
\usage{
excludeIdenticalFactors(dtPop)
}
\arguments{
\item{dtPop}{A data.table containing population data.}
}
\value{
A data.table with identical factor columns excluded.
}
\description{
This helper function checks for and excludes identical factor columns in the data.table.
}
