% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-exportResult.R
\name{exportIndividualResultsToPkml}
\alias{exportIndividualResultsToPkml}
\title{Export Individual Results to PKML}
\usage{
exportIndividualResultsToPkml(
  projectConfiguration,
  scenarioList,
  dtList,
  outputDir,
  individualId
)
}
\arguments{
\item{projectConfiguration}{A ProjectConfiguration object containing project configuration details, including paths for saving PKML files.}

\item{scenarioList}{A list of scenarios, each containing simulation parameters.}

\item{dtList}{A list of data.tables containing prior, start values, and mapped paths.}

\item{outputDir}{A string representing the directory where the PKML files will be saved.}

\item{individualId}{A string representing the ID of the individual whose results will be exported.}
}
\description{
This function exports individual results to a PKML file for a specified individual ID across scenarios.
}
