% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-exportResult.R
\name{exportIndividualValuesToConfigTable}
\alias{exportIndividualValuesToConfigTable}
\title{Export Individual Values to Configuration Table}
\usage{
exportIndividualValuesToConfigTable(projectConfiguration, scenarioList, dtList)
}
\arguments{
\item{projectConfiguration}{A ProjectConfiguration object containing project configuration details, including the path to the Excel file.}

\item{scenarioList}{A list of scenarios, each containing simulation parameters.}

\item{dtList}{A data.table containing BMLM configuration tables.}
}
\description{
This function exports individual values from a given data table to a specified
configuration table in an Excel workbook.
}
