% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-exportResult.R
\name{exportModelParametersToConfigTables}
\alias{exportModelParametersToConfigTables}
\title{Export Global and HyperParameter Parameters to Configuration Tables}
\usage{
exportModelParametersToConfigTables(
  projectConfiguration,
  dtList,
  sheetName,
  overwrite = FALSE
)
}
\arguments{
\item{projectConfiguration}{A ProjectConfiguration object containing project configuration details, including the path to the model parameters file.}

\item{dtList}{A list of data.tables containing the prior values.}

\item{sheetName}{A string representing the name of exported sheet.}

\item{overwrite}{A boolean indicating whether to overwrite an existing sheet.}
}
\description{
This function exports global parameters from the provided data table to a new sheet in the model parameters Excel file.
}
