% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-exportResult.R
\name{exportOptimizedPopulation}
\alias{exportOptimizedPopulation}
\title{Export Optimized Population}
\usage{
exportOptimizedPopulation(
  projectConfiguration,
  dtList,
  scenarioList,
  runName,
  addAsNewScenarios = TRUE,
  overwrite = FALSE
)
}
\arguments{
\item{projectConfiguration}{A ProjectConfiguration object containing project configuration details, including paths for saving population files.}

\item{dtList}{A list of data.tables containing the prior, start values, and mapped paths.}

\item{scenarioList}{A list of scenarios, each containing simulation parameters.}

\item{runName}{A string representing the name of the run.}

\item{addAsNewScenarios}{A boolean indicating whether to add new scenarios to the workbook.}

\item{overwrite}{A boolean indicating whether to overwrite existing files.}
}
\description{
This function exports optimized population data to CSV files for each scenario in the scenario list.
}
\concept{export}
