% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-exportResult.R
\name{exportPopulationWithVariability}
\alias{exportPopulationWithVariability}
\title{Export Population with Variability}
\usage{
exportPopulationWithVariability(
  projectConfiguration,
  populationName,
  variabilitySheetName,
  newName = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{projectConfiguration}{A ProjectConfiguration object containing project configuration details,
including the path to the populations folder and populations file.}

\item{populationName}{A string representing the name of the population (without .csv extension).}

\item{variabilitySheetName}{A string representing the name of the variability sheet in the
Populations.xlsx file.}

\item{newName}{An optional string for the new population name. If NULL, defaults to
paste(populationName, variabilitySheetName, sep = '_').}

\item{overwrite}{A logical indicating whether to overwrite an existing population file with the same name.
Default is FALSE.}
}
\value{
NULL (invisible). The function saves a new population CSV file.
}
\description{
This function loads an existing population CSV file and generates new random parameter values
according to the distributions defined in a variability sheet. Parameters with the same
parameter group are strictly correlated.
}
\seealso{
Other export: 
\code{\link{exportOptimizedPopulation}()}
}
\concept{export}
