% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-config.R
\name{extractIdentificationParameter}
\alias{extractIdentificationParameter}
\title{Title}
\usage{
extractIdentificationParameter(linkedParameter, headers)
}
\arguments{
\item{headers}{}
}
\description{
Title
}
