% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotParameter.R
\name{generateParameterDistributionPlot}
\alias{generateParameterDistributionPlot}
\title{Generate Parameter Distribution Plot}
\usage{
generateParameterDistributionPlot(
  dtValuesSubset,
  hyperParameterSubset,
  titeltxt,
  xScale,
  colorScalingVector,
  nCols
)
}
\arguments{
\item{dtValuesSubset}{A data.table containing the subset of values to plot.}

\item{hyperParameterSubset}{A data.table containing hyperparameter information.}

\item{titeltxt}{A string to include in the plot title.}

\item{xScale}{A character string specifying the scale of the x-axis.}

\item{colorScalingVector}{A named vector of colors for different statuses.}

\item{nCols}{An integer specifying the number of columns for faceting.}
}
\value{
A ggplot object representing the parameter distribution.
}
\description{
This function generates a ggplot object for parameter distributions.
}
\keyword{internal}
