% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-distribution.R
\name{getAllDistributions}
\alias{getAllDistributions}
\title{Get All Distributions}
\usage{
getAllDistributions()
}
\value{
A vector of unique distribution names.
}
\description{
This function retrieves all distributions available for prior definitions and hyperprameter.
}
\seealso{
Other distribution: 
\code{\link{computeStatFunction}()},
\code{\link{getDistributionParameters}()},
\code{\link{lognormal_geomean}}
}
\concept{distribution}
