% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotConvergence.R
\name{getConvergenceColumnHeaders}
\alias{getConvergenceColumnHeaders}
\title{Get Column Headers for Display Variables}
\usage{
getConvergenceColumnHeaders(dt, displayVariablesIndx)
}
\arguments{
\item{dt}{A data.table containing convergence data.}

\item{displayVariablesIndx}{An integer vector specifying which variables to display.}
}
\value{
A named character vector of column headers for the display variables.
}
\description{
This function retrieves the column headers for the specified display variable indices.
}
\keyword{internal}
