% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-dataPreparation.R
\name{getCurrentConfigTable}
\alias{getCurrentConfigTable}
\title{Get Current Configuration Table}
\usage{
getCurrentConfigTable(
  projectConfiguration,
  dtList,
  sheetName = c("Prior", "IndividualStartValues")
)
}
\arguments{
\item{projectConfiguration}{A ProjectConfiguration object containing project configuration details, including paths to Excel files.}

\item{dtList}{A list of data.tables containing the values.}

\item{sheetName}{A character string specifying the name of the sheet to retrieve data from. Options are 'Prior' or 'IndividualStartValues'.}
}
\value{
A data.table containing the current configuration values.
}
\description{
This function retrieves the current configuration table from the specified sheet in the Excel workbook.
}
\seealso{
Other configuration: 
\code{\link{addBMLMPConfiguration}()},
\code{\link{addBMLMWorkflowTemplate}()},
\code{\link{configurePriors}()},
\code{\link{convertVersionBMLM}()},
\code{\link{loadListsForRun}()},
\code{\link{readIdentificationParameterFromSnapshot}()}
}
\concept{configuration}
