% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-scenarioSimulation.R
\name{getDefaultPopulationLine}
\alias{getDefaultPopulationLine}
\title{Title}
\usage{
getDefaultPopulationLine(projectConfig, dtScenarioG, params)
}
\arguments{
\item{params}{}
}
\description{
Title
}
