% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-distribution.R
\name{getDistributionRow}
\alias{getDistributionRow}
\title{Get Distribution Row}
\usage{
getDistributionRow(distributionName, distributionParameter)
}
\arguments{
\item{distributionName}{A string representing the name of the distribution.}

\item{distributionParameter}{A string representing the parameter of the distribution.}
}
\value{
A data.table containing the distribution row.
}
\description{
This function retrieves a row from the distribution table based on the distribution name and parameter.
}
\keyword{internal}
