% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotParameter.R
\name{getFacetToPlotList}
\alias{getFacetToPlotList}
\title{Get Facet to Plot List}
\usage{
getFacetToPlotList(labelVector, nCols, nRows)
}
\arguments{
\item{labelVector}{A character or factor vector containing the facet labels.}

\item{nCols}{An integer specifying the number of columns in each plot.}

\item{nRows}{An integer specifying the number of rows in each plot.}
}
\value{
A list where each element is a vector of facets that should be plotted
together in a single plot.
}
\description{
This function generates a list of facets to be plotted based on the provided
label vector and the specified number of columns and rows for each plot.
}
\keyword{internal}
