% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-BMLMParameter.R
\name{getHyperParameter}
\alias{getHyperParameter}
\title{Get Hyper Parameters}
\usage{
getHyperParameter(dtPrior, dtDefinition)
}
\arguments{
\item{dtPrior}{A data.table containing prior definitions.}

\item{dtDefinition}{A data.table containing parameter definitions.}
}
\value{
A data.table containing hyperparameters and their distributions.
}
\description{
This function retrieves hyperparameters from the prior definition.
}
