% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-BMLMParameter.R
\name{getIndividualStartValues}
\alias{getIndividualStartValues}
\title{Get Individual Start Values}
\usage{
getIndividualStartValues(
  projectConfiguration,
  dtHyperParameter,
  dtDefinition,
  seed
)
}
\arguments{
\item{projectConfiguration}{An object of class ProjectConfiguration containing project configuration details, including the path to the BMLM configuration file.}

\item{dtHyperParameter}{A data.table containing hyperparameters and their distributions.}

\item{dtDefinition}{A data.table containing parameter definitions.}

\item{seed}{A numeric seed for random number generation.}
}
\value{
A data.table with updated start values.
}
\description{
This function reads individual start values from an Excel file and adjusts them based on hyperparameters.
}
