% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-BMLMParameter.R
\name{getLikelihoodForIndividualGroup}
\alias{getLikelihoodForIndividualGroup}
\title{Get Likelihood for Individual Group}
\usage{
getLikelihoodForIndividualGroup(indGroup, dtHyperParameter)
}
\arguments{
\item{indGroup}{A data.table containing a subset of start values for a specific individual group.}

\item{dtHyperParameter}{A data.table containing hyperparameters and their distributions.}
}
\value{
A numeric value representing the total log likelihood for the individual group.
}
\description{
This function calculates the likelihood for a specific group of individuals based on hyperparameters.
}
\keyword{internal}
