% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-BMLMParameter.R
\name{getLikelihoodHyperParameter}
\alias{getLikelihoodHyperParameter}
\title{Get Likelihood Hyperparameter}
\usage{
getLikelihoodHyperParameter(dtStartValues, dtHyperParameter)
}
\arguments{
\item{dtStartValues}{A data.table containing start values.}

\item{dtHyperParameter}{A data.table containing hyperparameters and their distributions.}
}
\value{
A numeric value representing the total log likelihood of hyperparameters.
}
\description{
This function calculates the likelihood of hyperparameters based on start values.
}
\keyword{internal}
