% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-loglikelihood.R
\name{getLikelihoodPriors}
\alias{getLikelihoodPriors}
\title{Calculate Likelihood Priors}
\usage{
getLikelihoodPriors(dtPrior)
}
\arguments{
\item{dtPrior}{A data.table containing prior information.}
}
\value{
The updated data.table with calculated probabilities.
}
\description{
This function calculates the likelihood priors for the given data.table.
}
\keyword{internal}
