% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-BMLMParameter.R
\name{getLikelihoodTimeProfiles}
\alias{getLikelihoodTimeProfiles}
\title{This function calculates the likelihood of observed data given simulated time profiles.}
\usage{
getLikelihoodTimeProfiles(
  scenarioList,
  dtPrior,
  dtStartValues,
  dataObservedForMatch,
  dtMappedPaths,
  simulationRunOptions
)
}
\arguments{
\item{scenarioList}{A list of scenarios for the optimization.}

\item{dtPrior}{A data.table containing prior values.}

\item{dtStartValues}{A data.table containing start values.}

\item{dataObservedForMatch}{A data.table containing observed data for matching.}

\item{dtMappedPaths}{A data.table containing mapped paths for parameters.}

\item{simulationRunOptions}{Optional parameters for running simulations.}
}
\value{
A numeric value representing the total log likelihood of the time profiles.
}
\description{
This function calculates the likelihood of observed data given simulated time profiles.
}
\keyword{internal}
