% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-loglikelihood.R
\name{getLogLikelihood}
\alias{getLogLikelihood}
\title{Get Log Likelihood}
\usage{
getLogLikelihood(
  params,
  scenarioList,
  dtList,
  simulationRunOptions,
  optimEnv = NULL
)
}
\arguments{
\item{params}{A numeric vector of parameters for the likelihood calculation.}

\item{scenarioList}{A list of scenarios for the optimization.}

\item{dtList}{A list containing various data.tables used in the optimization process.}

\item{simulationRunOptions}{Optional parameters for running simulations.}
}
\value{
A numeric value representing the negative log likelihood.
}
\description{
This function calculates the log likelihood based on the provided parameters and observed data.
}
\keyword{internal}
