% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-scaling.R
\name{getParams}
\alias{getParams}
\title{Prepare Input Data}
\usage{
getParams(
  dtPrior,
  dtStartValues,
  scalingMethod,
  valueColumn = c("value", "startValue"),
  optimizationGroup = c("both", "external", "internal")
)
}
\arguments{
\item{dtPrior}{A data.table containing prior values}

\item{dtStartValues}{A data.table containing start values}

\item{scalingMethod}{A character string specifying the scaling method.}

\item{valueColumn}{Name of column with value of interest either 'value' or 'startValue'.}

\item{optimizationGroup}{A character string indicating the optimization group;
must be one of 'both', 'external', or 'internal'.}
}
\value{
A data.table with combined and transformed input data.
}
\description{
This function prepares the input data for the L-BFGS-B algorithm by merging and transforming
the provided data tables. It handles log transformations based on specified conditions.
}
\keyword{internal}
