% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-startBMLMOptimization.R
\name{getPredictionsForScenario}
\alias{getPredictionsForScenario}
\title{Get Predictions for Scenarios}
\usage{
getPredictionsForScenario(
  scenarioResult,
  scenarioName,
  dataObservedForMatch,
  aggregationFun = NULL,
  identifier = c("outputPath", "individualId")
)
}
\arguments{
\item{scenarioResult}{An ScenarioResult object  containingsimulated results.}

\item{dataObservedForMatch}{A data.table containing observed data, which must include the required columns:
'scenario', 'outputPathId', 'yValues', and 'yUnit'.}

\item{aggregationFun}{A function for aggregation (optional). If provided, it will be used to aggregate the simulated results.}

\item{identifier}{A character vector of identifiers for matching, defaults to c("outputPath", "individualId").}
}
\value{
A data.table containing the predicted values for each scenario, along with the scenario name and other relevant identifiers.
}
\description{
This function generates predictions for a set of scenarios based on observed data.
}
\details{
The function validates the input names, loops through each scenario to generate predictions, and combines all results into a single data.table.
It also handles individual matching if 'ObservedIndividualId' is present in the scenario results.
}
\seealso{
Other data-preparation: 
\code{\link{prepareDataForMatch}()},
\code{\link{preparePopulationForCorrelationCheck}()}
}
\concept{data-preparation}
