% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-BMLMParameter.R
\name{getPriorDefinition}
\alias{getPriorDefinition}
\title{Get Prior Definition}
\usage{
getPriorDefinition(projectConfiguration, dtDefinition)
}
\arguments{
\item{projectConfiguration}{An object of class ProjectConfiguration containing project configuration details, including the path to the BMLM configuration file.}

\item{dtDefinition}{A data.table containing parameter definitions.}
}
\value{
A data.table with prior definitions.
}
\description{
This function retrieves prior definitions from the configuration file.
}
