% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-scaling.R
\name{getScaleFunction}
\alias{getScaleFunction}
\title{Get Scaling Function}
\usage{
getScaleFunction(scalingMethod)
}
\arguments{
\item{scalingMethod}{A character string specifying the scaling method.
Must be one of "unbounded" or "bounded".}
}
\value{
A function that performs the specified scaling operation.
}
\description{
This function returns the appropriate scaling function based on the specified scaling method.
}
\keyword{internal}
