% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-scaling.R
\name{getUnscaleFunction}
\alias{getUnscaleFunction}
\title{This function returns the appropriate inverse scaling function based on the specified scaling method.}
\usage{
getUnscaleFunction(scalingMethod)
}
\arguments{
\item{scalingMethod}{A character string specifying the scaling method.
Must be one of "unbounded" or "bounded".}
}
\value{
A function that performs the specified inverse scaling operation.
}
\description{
This function returns the appropriate inverse scaling function based on the specified scaling method.
}
\keyword{internal}
