% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-startBMLMOptimization.R
\name{initializeOptimEnv}
\alias{initializeOptimEnv}
\title{Initialize Optimization Environment}
\usage{
initializeOptimEnv(
  dtList,
  failValue = Inf,
  bestValue = Inf,
  NAcounter = 0,
  iteration = 0,
  params = c()
)
}
\arguments{
\item{dtList}{A list containing data tables for optimization.}

\item{failValue}{A numeric value to set if evaluation of the objective function fails.}
}
\value{
A new environment containing initialized optimization variables.
}
\description{
Initializes the optimization environment by extracting relevant values from
Initialize Optimization Environment
}
\details{
Initializes the optimization environment by extracting relevant values from
the provided data table list.
}
\keyword{internal}
