% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-startBMLMOptimization.R
\name{initializeOptimization}
\alias{initializeOptimization}
\title{Initialize Optimization}
\usage{
initializeOptimization(restartAtBestValues, outputDir, dtList)
}
\arguments{
\item{restartAtBestValues}{A logical indicating whether to restart from the best values found.}

\item{outputDir}{A character string representing the path to the output directory.}

\item{dtList}{A list containing data tables for optimization.}
}
\value{
A numeric value representing the initialized best value.
}
\description{
Initializes optimization variables and loads previous results if restarting from the best values.
}
\keyword{internal}
