% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-dataPreparation.R
\name{inverseTransformParams}
\alias{inverseTransformParams}
\title{Inverse Transform Parameters}
\usage{
inverseTransformParams(param, minValue, maxValue, scaling)
}
\arguments{
\item{param}{A numeric value of parameters to be transformed back.}

\item{minValue}{A numeric value of minimum values.}

\item{maxValue}{A numeric value of maximum values.}

\item{scaling}{A value indicating the scaling type for each parameter.}
}
\value{
A numeric vector of original values.
}
\description{
This function applies the inverse transformation to the parameters to retrieve original values.
}
\keyword{internal}
