% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-dataPreparation.R
\name{loadListsForRun}
\alias{loadListsForRun}
\title{Load Lists for Run}
\usage{
loadListsForRun(outputDir, runName)
}
\arguments{
\item{runName}{A character string representing the name of the run, which is used to
construct the output directory path.}

\item{projectConfiguration}{An object of class projectConfiguration containing configuration settings, including
the output folder path where the CSV files are located.}
}
\value{
A list of data.tables, each corresponding to one of the CSV files loaded. The
list elements are named after the CSV files (without the '.csv' extension).
}
\description{
This function loads various CSV files from a specified output directory for a given name of a BLMLM run,
processes them into data.tables, and returns them as a list.
}
\seealso{
Other configuration: 
\code{\link{addBMLMPConfiguration}()},
\code{\link{addBMLMWorkflowTemplate}()},
\code{\link{configurePriors}()},
\code{\link{convertVersionBMLM}()},
\code{\link{getCurrentConfigTable}()},
\code{\link{readIdentificationParameterFromSnapshot}()}
}
\concept{configuration}
