% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-exportResult.R
\name{loadOrCreateSheetData}
\alias{loadOrCreateSheetData}
\title{Load or Create Sheet Data}
\usage{
loadOrCreateSheetData(wb, sheetName)
}
\arguments{
\item{wb}{An open workbook object.}

\item{sheetName}{A string representing the name of the sheet to load.}
}
\value{
A data.table containing the data from the specified sheet or an empty data.table if the sheet does not exist.
}
\description{
This function loads data from a specified sheet in an Excel workbook or creates an empty data.table if the sheet does not exist.
}
\keyword{internal}
