% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-config.R
\name{loadPriorData}
\alias{loadPriorData}
\title{Load Prior Data}
\usage{
loadPriorData(wb, overwrite)
}
\arguments{
\item{wb}{An open workbook object from which to read the prior data.}

\item{overwrite}{A logical indicating whether to overwrite existing data.
Default is FALSE.}
}
\value{
A data.table containing the prior data.
}
\description{
This function loads the prior data from the specified workbook and handles
the logic for overwriting the existing data if necessary.
}
\keyword{internal}
