% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-config.R
\name{loadStartValuesHeader}
\alias{loadStartValuesHeader}
\title{Load Start Values Header}
\usage{
loadStartValuesHeader(wb, overwrite)
}
\arguments{
\item{wb}{An open workbook object from which to read the start values header.}

\item{overwrite}{A logical indicating whether to overwrite existing data.
Default is FALSE.}
}
\value{
A data.table containing the start values header.
}
\description{
This function loads the start values header from the specified workbook and
handles the logic for overwriting the existing header if necessary.
}
