% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-startBMLMOptimization.R
\name{logAndDisplayOptimization}
\alias{logAndDisplayOptimization}
\title{Log and Display Optimization Messages}
\usage{
logAndDisplayOptimization(
  message,
  outputDir,
  quiet = TRUE,
  prependTimestamp = TRUE
)
}
\arguments{
\item{message}{A character string containing the message to log.}

\item{outputDir}{A character string specifying the directory where the log file will be saved.}

\item{quiet}{A logical value indicating whether to suppress console output. Defaults to TRUE.}

\item{prependTimestamp}{A logical value indicating whether to prepend the current timestamp to the message. Defaults to TRUE.}
}
\value{
NULL (invisible) This function is called for its side effects (logging and printing).
}
\description{
This function logs messages related to the optimization process to a specified file and optionally prints them to the console.
}
\keyword{internal}
