% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-startBMLMOptimization.R
\name{logFunctionCall}
\alias{logFunctionCall}
\title{Log Function Call}
\usage{
logFunctionCall(
  logFile,
  projectConfiguration,
  scenarioList,
  dataObservedExpr,
  seed,
  method,
  controlExpr,
  restartAtBestValues,
  failValue,
  ...
)
}
\arguments{
\item{logFile}{A character string representing the path to the log file.}

\item{projectConfiguration}{A list containing project configuration details.}

\item{scenarioList}{A list of scenarios to be optimized.}

\item{dataObservedExpr}{An expression representing the observed data; useful for logging purposes.}

\item{seed}{An integer for random number generation.}

\item{method}{A character string specifying the optimization method.}

\item{controlExpr}{A list of control parameters for the optimization function.}

\item{restartAtBestValues}{A logical indicating whether to restart from the best values found.}

\item{failValue}{A numeric value to set if evaluation of the objective function fails.}

\item{...}{Additional arguments passed to the optimization function.}
}
\description{
Logs the details of the function call to the specified log file.
}
\keyword{internal}
