% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-startBMLMOptimization.R
\name{logMessage}
\alias{logMessage}
\title{Log Message}
\usage{
logMessage(logFile, message)
}
\arguments{
\item{logFile}{A character string representing the path to the log file.}

\item{message}{A character string containing the message to log.}
}
\description{
Logs a message to the specified log file with a timestamp.
}
\keyword{internal}
