% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-startBMLMOptimization.R
\name{manageOutputDirectory}
\alias{manageOutputDirectory}
\title{Manage Output Directory}
\usage{
manageOutputDirectory(baseDir, runName, restartAtBestValues)
}
\arguments{
\item{baseDir}{A character string specifying the base directory where the output
folder will be created or managed.}

\item{runName}{A character string specifying the name of the run, which will
be used to create the output folder.}

\item{restartAtBestValues}{A logical value indicating whether to restart at the
best values. If FALSE and the directory exists, a new directory will be created
with a timestamp appended to the run name.}
}
\value{
A character string containing the path to the output directory, or NULL
if the operation is cancelled by the user.
}
\description{
This function manages the output directory for a specific run.
If \code{restartAtBestValues} is FALSE and the directory exists, a new directory
will be created with a timestamp appended to the run name.
}
\keyword{internal}
