% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\docType{data}
\name{messages}
\alias{messages}
\title{List of functions and strings used to signal error and warning messages}
\format{
An object of class \code{list} of length 37.
}
\usage{
messages
}
\description{
This module contains all error and warning messages used throughout the package.
Messages are organized as functions that return formatted strings.
}
\keyword{datasets}
