% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-template.R
\name{openBMLMWorkflowTemplate}
\alias{openBMLMWorkflowTemplate}
\title{Opens the workflow template as new document}
\usage{
openBMLMWorkflowTemplate()
}
\description{
Per default the workflow template of the package is loaded.
To customize it save the file with name "template_workflow.R" in your template directory and
set the path to the template via option \verb{options(OSPSuite.RF.PathForWorkflowTemplate = 'myTemplateDirectory'})
}
