% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ospsuit.bmlm-package.R
\docType{package}
\name{ospsuite.bmlm-package}
\alias{ospsuite.bmlm}
\alias{ospsuite.bmlm-package}
\title{ospsuite.bmlm: Library for multi level parameter identification}
\description{
This library provides a toolbox to identify parameters in multi level models. It uses a bayesian approach.
}
\author{
\strong{Maintainer}: Katrin Coboeken \email{katrin.coboeken@bayer.com}

Other contributors:
\itemize{
  \item Open-Systems-Pharmacology Community [copyright holder]
}

}
\keyword{internal}
