% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-enum.R
\docType{data}
\name{PARAMETERTYPE}
\alias{PARAMETERTYPE}
\title{Parameter types}
\format{
An object of class \code{list} of length 4.
}
\usage{
PARAMETERTYPE
}
\description{
Parameter types
}
\seealso{
Other enum helpers: 
\code{\link{ERRORMODEL}},
\code{\link{RUNSTATUS}},
\code{\link{SCALING}},
\code{\link{SCALINGMETHOD}}
}
\concept{enum helpers}
\keyword{datasets}
