% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotParameter.R
\name{plotBestVsStartParameter}
\alias{plotBestVsStartParameter}
\title{Plot Best Values vs Start Values}
\usage{
plotBestVsStartParameter(
  dtList,
  statusList,
  nCols = 2,
  nRows = 3,
  xyScale = unlist(SCALING),
  parameterFilter = NULL,
  titeltxt = NULL,
  ...
)
}
\arguments{
\item{dtList}{A list containing prior and start values data tables.}

\item{statusList}{A list containing the current and best parameter values.}

\item{nCols}{An integer specifying the number of columns for faceting. Default is 2.}

\item{nRows}{An integer specifying the maximum number of rows for faceting. Default is 3.}

\item{xyScale}{A character string specifying the scale of the x- and y-axis. Default is 'log'.}

\item{parameterFilter}{A character vector for filtering parameter names. Default is NULL.}

\item{titeltxt}{A string to include in the plot title.}
}
\value{
An invisible ggplot object visualizing the best values vs start values.
}
\description{
This function generates a plot comparing the best values against the start values.
}
\seealso{
Other plotting: 
\code{\link{defaultColorsParameterPlots}()},
\code{\link{plotConvergence}()},
\code{\link{plotCorrelations}()},
\code{\link{plotDistributions}()},
\code{\link{plotParameterLimits}()},
\code{\link{plotParameterValuesVsPrior}()},
\code{\link{plotPredictedVsObserved}()},
\code{\link{plotPredictedVsTime}()},
\code{\link{plotResidualsAsHistogram}()},
\code{\link{plotResidualsAsQQ}()},
\code{\link{plotResidualsDistribution}()},
\code{\link{plotResidualsVsTime}()}
}
\concept{plotting}
