% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotParameter.R
\name{plotParameterValuesVsPrior}
\alias{plotParameterValuesVsPrior}
\title{Create and Print Parameter Values vs Prior Plot}
\usage{
plotParameterValuesVsPrior(
  dtList,
  statusList,
  xScale = unlist(SCALING),
  titeltxt = NULL,
  colorScalingVector = defaultColorsParameterPlots()
)
}
\arguments{
\item{dtList}{A list containing prior and start values data tables.}

\item{statusList}{A list containing the current and best parameter values.}

\item{xScale}{A character string specifying the scale of the x-axis. Default is 'log'.}

\item{titeltxt}{A string to include in the plot title.}

\item{colorScalingVector}{A named vector of colors for different statuses.
Default is set by \code{defaultColorsParameterPlots()}.}
}
\value{
An invisible ggplot object visualizing the parameter values vs prior.
}
\description{
This function generates a plot comparing parameter values against their prior distributions.
}
