% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPredicted.R
\name{plotPredictedVsTime}
\alias{plotPredictedVsTime}
\title{Create and Print Predicted vs Time Plots}
\usage{
plotPredictedVsTime(
  dtRes,
  yScale = unlist(SCALING),
  nCols = 4,
  titeltxt = NULL,
  ...
)
}
\arguments{
\item{dtRes}{A data frame containing the data to be plotted. It should include columns for
outputPathId, scenarioName, individualId, predicted, yValues, and yUnit.}

\item{yScale}{A character string specifying the scale for the y-axis (default is "log").}

\item{nCols}{An integer specifying the number of columns for the facet wrap (default is 4).}

\item{titeltxt}{A string to include in the plot title.}

\item{...}{Additional arguments passed on to ospsuite.plots::plotTimeProfile.}
}
\value{
A list of ggplot objects for each unique outputPathId and scenarioName combination.
}
\description{
This function generates plots to compare predicted values against time.
}
