% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPredicted.R
\name{plotResidualLoop}
\alias{plotResidualLoop}
\title{Plot Residual Loop Function}
\usage{
plotResidualLoop(
  dtRes,
  plotFunction,
  nCols = 2,
  titeltxt = NULL,
  excludeCensored = FALSE,
  ...
)
}
\arguments{
\item{dtRes}{A data frame containing the data to be plotted. It should include columns for
scenario, outputPathId, group, and the residuals.}

\item{plotFunction}{A function that defines the specific plotting logic for each residual plot.}

\item{nCols}{An integer specifying the number of columns for the facet wrap (default is 2).}

\item{titeltxt}{A string to include in the plot title.}

\item{excludeCensored}{A logical value indicating whether to exclude censored data (default is FALSE).}

\item{...}{Additional arguments passed to the specific plotting function.}
}
\value{
NULL This function does not return a value; it prints the plots directly.
}
\description{
This function encapsulates the common logic for generating residual plots.
}
\keyword{internal}
