% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPredicted.R
\name{plotResidualsAsHistogram}
\alias{plotResidualsAsHistogram}
\title{Create and Print Residuals as Histogram Plots}
\usage{
plotResidualsAsHistogram(filteredData, ...)
}
\arguments{
\item{filteredData}{A data frame containing the data to be plotted. It should include columns for
scenario, outputPathId, group, and the residuals.}

\item{...}{Additional arguments passed on to ospsuite.plots::plotHistogram.}
}
\value{
A ggplot object for the histogram plot.
}
\description{
This function generates histogram plots of the residuals to visualize their distribution for a single outputPathId.
}
