% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPredicted.R
\name{plotResidualsAsQQ}
\alias{plotResidualsAsQQ}
\title{Create and Print Residuals as QQ Plot}
\usage{
plotResidualsAsQQ(filteredData, ...)
}
\arguments{
\item{filteredData}{A data frame containing the data to be plotted. It should include columns for
scenario, outputPathId, group, and the residuals.}

\item{...}{Additional arguments passed on to ospsuite.plots::plotQQ.}
}
\value{
A ggplot object for the QQ plot.
}
\description{
This function generates a QQ plot to assess the normality of the residuals for a single outputPathId.
}
\seealso{
Other plotting: 
\code{\link{defaultColorsParameterPlots}()},
\code{\link{plotBestVsStartParameter}()},
\code{\link{plotConvergence}()},
\code{\link{plotCorrelations}()},
\code{\link{plotDistributions}()},
\code{\link{plotParameterLimits}()},
\code{\link{plotParameterValuesVsPrior}()},
\code{\link{plotPredictedVsObserved}()},
\code{\link{plotPredictedVsTime}()},
\code{\link{plotResidualsAsHistogram}()},
\code{\link{plotResidualsDistribution}()},
\code{\link{plotResidualsVsTime}()}
}
\concept{plotting}
