% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPredicted.R
\name{plotResidualsDistribution}
\alias{plotResidualsDistribution}
\title{Plot Residual Distribution}
\usage{
plotResidualsDistribution(filteredData, ...)
}
\arguments{
\item{filteredData}{A data.table containing the residuals and other relevant data.}

\item{...}{Additional arguments passed to the plotting function.}
}
\value{
A ggplot object visualizing the residual distribution.
}
\description{
This function creates a plot to visualize the distribution of residuals from the model predictions.
}
\seealso{
Other plotting: 
\code{\link{defaultColorsParameterPlots}()},
\code{\link{plotBestVsStartParameter}()},
\code{\link{plotConvergence}()},
\code{\link{plotCorrelations}()},
\code{\link{plotDistributions}()},
\code{\link{plotParameterLimits}()},
\code{\link{plotParameterValuesVsPrior}()},
\code{\link{plotPredictedVsObserved}()},
\code{\link{plotPredictedVsTime}()},
\code{\link{plotResidualsAsHistogram}()},
\code{\link{plotResidualsAsQQ}()},
\code{\link{plotResidualsVsTime}()}
}
\concept{plotting}
