% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotParameter.R
\name{prepareDataForDistributionPlot}
\alias{prepareDataForDistributionPlot}
\title{Prepare Data for Distribution Plot}
\usage{
prepareDataForDistributionPlot(
  dtList,
  statusList,
  parameterFilter,
  zoomOnData,
  xScale
)
}
\arguments{
\item{dtList}{A list containing prior and start values data tables.}

\item{statusList}{A list containing the current and best parameter values.}

\item{zoomOnData}{A logical value indicating whether to zoom in on data. Default is FALSE.}

\item{xScale}{display scale of x-axis.}
}
\value{
A data.table containing prepared data for distribution plots.
}
\description{
This function prepares data for distribution plots.
}
\keyword{internal}
