% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotParameter.R
\name{prepareDataForParameterLimits}
\alias{prepareDataForParameterLimits}
\title{Prepare Data for Parameter Limits}
\usage{
prepareDataForParameterLimits(dtList, statusList)
}
\arguments{
\item{dtList}{A list containing prior and start values data tables.}

\item{statusList}{A list containing the current and best parameter values.}
}
\value{
A list containing processed data for individuals and globals.
}
\description{
This function prepares data for plotting parameter limits.
}
\keyword{internal}
