% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-dataPreparation.R
\name{prepareInputData}
\alias{prepareInputData}
\title{Prepare Input Data}
\usage{
prepareInputData(
  dtPrior,
  dtStartValues,
  valueColumn = c("value", "startValue")
)
}
\arguments{
\item{dtPrior}{A data.table containing prior values}

\item{dtStartValues}{A data.table containing start values}

\item{valueColumn}{Name of column with value of interest either 'value' or 'startValue'.}
}
\value{
A data.table with combined and transformed input data.
}
\description{
This function prepares the input data for the L-BFGS-B algorithm by merging and transforming
the provided data tables. It handles log transformations based on specified conditions.
}
\keyword{internal}
