% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-configuration.R
\docType{class}
\name{ProjectConfigurationBMLM}
\alias{ProjectConfigurationBMLM}
\title{ProjectConfiguration}
\description{
An object storing configuration used project-wide
}
\section{Super classes}{
\code{\link[ospsuite.utils:Printable]{ospsuite.utils::Printable}} -> \code{\link[esqlabsR:ProjectConfiguration]{esqlabsR::ProjectConfiguration}} -> \code{ProjectConfiguration}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{BMLMConfigurationFile}}{Path to the file containing BMLM Identification
Initialize}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ProjectConfiguration-new}{\code{ProjectConfigurationBMLM$new()}}
\item \href{#method-ProjectConfiguration-clone}{\code{ProjectConfigurationBMLM$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="esqlabsR" data-topic="ProjectConfiguration" data-id="print"><a href='../../esqlabsR/html/ProjectConfiguration.html#method-ProjectConfiguration-print'><code>esqlabsR::ProjectConfiguration$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="esqlabsR" data-topic="ProjectConfiguration" data-id="save"><a href='../../esqlabsR/html/ProjectConfiguration.html#method-ProjectConfiguration-save'><code>esqlabsR::ProjectConfiguration$save()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectConfiguration-new"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectConfiguration-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectConfigurationBMLM$new(projectConfigurationFilePath = character())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{projectConfigurationFilePath}}{A string representing the path to the
project configuration file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectConfiguration-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectConfiguration-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectConfigurationBMLM$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
