% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-BMLMParameter.R
\name{randomizeIndividualStartValues}
\alias{randomizeIndividualStartValues}
\title{Randomize Start Values}
\usage{
randomizeIndividualStartValues(indGroup, dtHyperParameter)
}
\arguments{
\item{indGroup}{A data.table containing a subset of start values for a specific individual group.}

\item{dtHyperParameter}{A data.table containing hyperparameters and their distributions.}
}
\value{
A data.table with updated start values for the individual group.
}
\description{
This function processes a single group of start values, merging with hyperparameters and generating new values.
}
