% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data.R
\name{readDataByDictionary}
\alias{readDataByDictionary}
\title{Title}
\usage{
readDataByDictionary(projectConfig, weightingModus = WEIGHTINGMODUS$allOne)
}
\arguments{
\item{weightingModus}{mod to set data
\itemize{
\item "allOne" each data point has the same weight $w_{i,j}(t) = 1 $
\item "weightPerIndividual" each individual has the same weight $w_{i,j}(t) = 1/N_i$, $N_i$ number of data points per individual
\item "weightPerOutput" each observer has the same weight $w_{i,j}(t) = 1/N_j$, $N_i$ number of data points per observer
\item "weightPerOutputPerIndividual" each observer per individual has the same weight $w_{i,j}(t) = 1/N_{i,j}$, $N_{i,j}$ number of data points per observer and individual
\item "weightPerOutputAndIndividual" each individual has the same weight and within a individual each observer has the same weight $w_{i,j}(t) = 1/N_{i,j}/N_J$, $N_{i,j}$ number of data points per observer and individual $N_J$ number of outputs per individual
}}
}
\description{
Title
}
