% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-config.R
\name{readIdentificationParameterFromSnapshot}
\alias{readIdentificationParameterFromSnapshot}
\title{Read Identification Parameter From Snapshot}
\usage{
readIdentificationParameterFromSnapshot(
  snapshotFile,
  nameOfParameterIdentification,
  projectConfiguration,
  overwrite = FALSE
)
}
\arguments{
\item{snapshotFile}{A string representing the path to the snapshot file. This file should be in JSON format and contain
parameter identifications that can be extracted and used to update the project configuration.}

\item{nameOfParameterIdentification}{A string representing the name of the parameter identification to be read from the snapshot.
This should correspond to a valid entry under \code{ParameterIdentifications} in the snapshot file.}

\item{projectConfiguration}{An object of class \code{ProjectConfigurationBMLM}, which contains configuration details for the project.
This object will be modified to incorporate the new identification parameters extracted from the snapshot.}

\item{overwrite}{A logical indicating whether to overwrite existing data in the ParameterDefinition sheet of the configuration file.
Default is FALSE. If TRUE, existing data will be replaced with the new values from the snapshot.}
}
\value{
NULL (invisible), as the function performs side effects by modifying the configuration file directly.
}
\description{
This function reads specified identification parameters from a snapshot file and updates the BMLM project configuration accordingly.
It performs checks to ensure the validity of the snapshot and the specified parameter identification.
}
\seealso{
Other configuration: 
\code{\link{addBMLMPConfiguration}()},
\code{\link{addBMLMWorkflowTemplate}()},
\code{\link{configurePriors}()},
\code{\link{convertVersionBMLM}()},
\code{\link{getCurrentConfigTable}()},
\code{\link{loadListsForRun}()}
}
\concept{configuration}
