% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-config.R
\name{readProjectConfiguration}
\alias{readProjectConfiguration}
\title{Title}
\usage{
readProjectConfiguration(configurationfile, properties = list())
}
\arguments{
\item{configurationfile}{'}
}
\description{
Title
}
