% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-startBMLMOptimization.R
\name{runInternalOptimization}
\alias{runInternalOptimization}
\title{Run Internal Optimization}
\usage{
runInternalOptimization(dtList, optimEnv, outputDir)
}
\arguments{
\item{dtList}{A list containing data tables relevant for the optimization process.}

\item{optimEnv}{An environment containing the current state variables for optimization.}
}
\value{
A list containing the results of the internal optimization, including the optimized parameters
and the value of the objective function.
}
\description{
Executes the internal optimization process using the specified parameters.
}
\details{
This function performs optimization on internal parameters based on the provided
data tables and the current optimization environment. It utilizes the \code{optim}
function to find the best internal parameters that maximize the log likelihood.

The internal optimization is performed using the BFGS method, which is suitable for
smooth optimization problems. The function retrieves initial parameter values and
evaluates the log likelihood to determine the best fit.
}
\keyword{internal}
