% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-enum.R
\docType{data}
\name{RUNSTATUS}
\alias{RUNSTATUS}
\title{Status of optimization Runs}
\format{
An object of class \code{list} of length 4.
}
\usage{
RUNSTATUS
}
\description{
Status of optimization Runs
}
\seealso{
Other enum helpers: 
\code{\link{ERRORMODEL}},
\code{\link{PARAMETERTYPE}},
\code{\link{SCALING}},
\code{\link{SCALINGMETHOD}}
}
\concept{enum helpers}
\keyword{datasets}
