% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-dataPreparation.R
\name{saveDataTablesAsCSV}
\alias{saveDataTablesAsCSV}
\title{Save Data Tables as CSV}
\usage{
saveDataTablesAsCSV(dtList, outputDir)
}
\arguments{
\item{dtList}{A list containing data tables to be saved.}

\item{outputDir}{A character string representing the path to the output directory.}
}
\description{
Saves the provided data tables to CSV files in the specified output directory.
}
\keyword{internal}
