% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-exportResult.R
\name{saveDataToWorkbook}
\alias{saveDataToWorkbook}
\title{Save Data to Workbook}
\usage{
saveDataToWorkbook(
  wb,
  sheetName,
  dt,
  templateSheet = "template_Ind",
  templateXlsx = "Individuals.xlsx"
)
}
\arguments{
\item{wb}{An open workbook object.}

\item{sheetName}{A string representing the name of the sheet to save data to.}

\item{dt}{A data.table containing the data to be saved.}

\item{templateSheet}{A string representing the name of the template sheet to use if the sheet does not exist.}

\item{templateXlsx}{A string representing the name of the template Excel file.}
}
\description{
This function saves a data.table to a specified sheet in an Excel workbook, creating the sheet from a template if it does not exist.
}
