% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-exportResult.R
\name{saveFinalValuesToTables}
\alias{saveFinalValuesToTables}
\title{Save Final Values to Tables}
\usage{
saveFinalValuesToTables(projectConfiguration, dtList)
}
\arguments{
\item{projectConfiguration}{A ProjectConfiguration object containing project configuration details, including paths to Excel files.}

\item{dtList}{A list of data.tables containing the final values to be saved.}
}
\value{
An updated workbook object.
}
\description{
This function saves final values from a provided data table to specified sheets in an Excel workbook.
}
\keyword{internal}
