% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-startBMLMOptimization.R
\name{saveOptimStatusIfNeeded}
\alias{saveOptimStatusIfNeeded}
\title{Save Optimization Status If Needed}
\usage{
saveOptimStatusIfNeeded(optimStatus, optimEnv, outputDir, interval)
}
\arguments{
\item{optimStatus}{A list containing the current optimization status.}

\item{optimEnv}{The optimization environment containing current state variables.}

\item{outputDir}{A character string representing the path to the output directory.}

\item{interval}{An integer specifying the interval for saving the status in seconds.}
}
\description{
Saves the current optimization status to a file if the specified time interval
has passed since the last save.
}
\keyword{internal}
