% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-scaling.R
\name{scaleToLogSig}
\alias{scaleToLogSig}
\title{Scale to Unbounded Values}
\usage{
scaleToLogSig(value, minValue, maxValue, scaling)
}
\arguments{
\item{value}{A numeric vector of values to be scaled.}

\item{minValue}{A numeric value representing the minimum bound.}

\item{maxValue}{A numeric value representing the maximum bound.}

\item{scaling}{A character string indicating the scaling method; currently supports "log".}
}
\value{
A numeric vector of scaled values.
}
\description{
This function scales input values to an unbounded range using the specified scaling method.
}
\keyword{internal}
