% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-scaling.R
\name{scaleWithinBounds}
\alias{scaleWithinBounds}
\title{Scale Within Specified Limits}
\usage{
scaleWithinBounds(value, minValue, maxValue, scaling)
}
\arguments{
\item{value}{A numeric vector of values to be scaled.}

\item{minValue}{A numeric value representing the minimum bound.}

\item{maxValue}{A numeric value representing the maximum bound.}

\item{scaling}{A character string indicating the scaling method; currently supports "log".}
}
\value{
A numeric vector of scaled values, or NA if the input is out of bounds.
}
\description{
This function scales input values to a range between 0 and 1, ensuring they are within specified limits.
}
\keyword{internal}
