% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-enum.R
\docType{data}
\name{SCALINGMETHOD}
\alias{SCALINGMETHOD}
\title{Methods to scale parameters}
\format{
An object of class \code{list} of length 2.
}
\usage{
SCALINGMETHOD
}
\description{
Methods to scale parameters
}
\seealso{
Other enum helpers: 
\code{\link{ERRORMODEL}},
\code{\link{PARAMETERTYPE}},
\code{\link{RUNSTATUS}},
\code{\link{SCALING}}
}
\concept{enum helpers}
\keyword{datasets}
