% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotConvergence.R
\name{selectIterations}
\alias{selectIterations}
\title{Select Points Based on Selection Mode}
\usage{
selectIterations(dtConvergenceList, nPointsAvailable, nPoints, selectionMode)
}
\arguments{
\item{nPointsAvailable}{An integer specifying the number of available points.}

\item{nPoints}{An integer specifying the number of points to select.}

\item{selectionMode}{A character string indicating the mode of selection for points.}

\item{dt}{A data.table containing convergence data.}
}
\value{
A data.table containing the selected points.
}
\description{
This function selects points from the convergence data based on the specified selection mode.
}
\keyword{internal}
