% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-loglikelihood.R
\name{setlogTruncationOffset}
\alias{setlogTruncationOffset}
\title{Calculate Log Truncation Offset for Hyperparameters}
\usage{
setlogTruncationOffset(
  dtPrior,
  dtStartValues,
  identifier = c("name", "categoricCovariate"),
  colsToKeep = c("hyperParameter", "value", "hyperDistribution", "scaling",
    "logTruncationOffset")
)
}
\arguments{
\item{dtPrior}{A data.table containing prior hyperparameter distributions. It should include columns for hyperparameters,
their values, and the distribution type.}

\item{dtStartValues}{A data.table containing the starting values for hyperparameters, including identifiers and their
respective minimum and maximum values.}

\item{identifier}{A character vector indicating the columns to be used as identifiers for grouping the hyperparameters.
Default is c('name', 'categoricCovariate').}

\item{colsToKeep}{A character vector indicating the columns to retain in the resulting data.table. Default includes
"hyperParameter", "value", "hyperDistribution", "scaling", and "logTruncationOffset".}
}
\value{
A data.table containing the hyperparameters, their values, and the computed log truncation offsets.
}
\description{
This function computes the log truncation offset for hyperparameters based on the provided prior distribution and start values.
It reduces the hyperparameter data to match the available data and calculates the lower and upper probabilities for the given
minimum and maximum values. The log truncation offset is then derived from these probabilities.
}
\keyword{internal}
