% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-scaling.R
\name{setParameterToTables}
\alias{setParameterToTables}
\title{Set Parameter to Tables}
\usage{
setParameterToTables(dtList, params, scalingMethod)
}
\arguments{
\item{dtList}{A list containing various data.tables used in the optimization process.}

\item{params}{A numeric vector of parameters for the likelihood calculation.}

\item{scalingMethod}{A character string specifying the scaling method.
Must be one of "logsig" or "hardBounds".}
}
\value{
A list containing the updated data.tables.
}
\description{
This function updates the parameter values in the provided data.tables based on the optimization results.
}
\keyword{internal}
