% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-dataPreparation.R
\name{transformToUnbounded}
\alias{transformToUnbounded}
\title{Transform to Unbounded Values}
\usage{
transformToUnbounded(value, minValue, maxValue, scaling)
}
\arguments{
\item{value}{A numeric value of values to be transformed.}

\item{minValue}{A numeric value of minimum values.}

\item{maxValue}{A numeric value of maximum values.}

\item{scaling}{A value indicating the scaling type for value.}
}
\value{
A numeric value of transformed values.
}
\description{
This function transforms parameters to unbounded values based on the scaling type.
}
\keyword{internal}
