% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-scaling.R
\name{unscaleFromBounds}
\alias{unscaleFromBounds}
\title{Inverse Scale from Specified Limits}
\usage{
unscaleFromBounds(param, minValue, maxValue, scaling)
}
\arguments{
\item{param}{A numeric vector of parameters to be unscaled.}

\item{minValue}{A numeric value representing the minimum bound.}

\item{maxValue}{A numeric value representing the maximum bound.}

\item{scaling}{A character string indicating the scaling method; currently supports "log".}
}
\value{
A numeric vector of unscaled values, or NA if the input is out of bounds.
}
\description{
This function inversely scales parameters from a range between 0 and 1 back to their original values.
}
\keyword{internal}
