% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-scaling.R
\name{unscaleFromLogSig}
\alias{unscaleFromLogSig}
\title{Inverse Scale from Unbounded Values}
\usage{
unscaleFromLogSig(param, minValue, maxValue, scaling)
}
\arguments{
\item{param}{A numeric vector of parameters to be unscaled.}

\item{minValue}{A numeric value representing the minimum bound.}

\item{maxValue}{A numeric value representing the maximum bound.}

\item{scaling}{A character string indicating the scaling method; currently supports "log".}
}
\value{
A numeric vector of unscaled values.
}
\description{
This function inversely scales parameters from an unbounded range back to their original values.
}
\keyword{internal}
