% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-startBMLMOptimization.R
\name{updateBestValueIfImproved}
\alias{updateBestValueIfImproved}
\title{Update Best Value If Improved}
\usage{
updateBestValueIfImproved(
  currentValue,
  optimEnv,
  optimStatus,
  outputDir,
  loglikelihoods
)
}
\arguments{
\item{currentValue}{A numeric value representing the current evaluation of the objective function.}

\item{optimEnv}{The optimization environment containing current state variables.}

\item{optimStatus}{A list containing the current optimization status.}

\item{outputDir}{A character string representing the path to the output directory.}

\item{loglikelihoods}{A vector of log likelihood values.}
}
\description{
Checks if the current value of the optimization is better than the best value
recorded and updates the best value if it is.
}
\keyword{internal}
