% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-config.R
\name{updateDefinitionDT}
\alias{updateDefinitionDT}
\title{Update Definition Data Table}
\usage{
updateDefinitionDT(linkedParameterDT, definitionDTHeader, wb)
}
\arguments{
\item{linkedParameterDT}{A data.table containing linked parameters, including their fixed status and values.}

\item{definitionDTHeader}{A data.table containing the header for the definition data.}

\item{wb}{A workbook object to write the updated data table.}
}
\value{
NULL (invisible), as the function performs side effects by writing to a workbook.
}
\description{
This helper function prepares and updates the definition data table by filtering and modifying the linked parameter data.
}
\keyword{internal}
