% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-config.R
\name{updateFixedParameters}
\alias{updateFixedParameters}
\title{Update Fixed Parameters}
\usage{
updateFixedParameters(
  linkedParameterDT,
  projectConfiguration,
  nameOfParameterIdentification
)
}
\arguments{
\item{linkedParameterDT}{A data.table containing linked parameters, including their fixed status and values.}

\item{projectConfiguration}{An object of class \code{ProjectConfigurationBMLM}, which contains configuration details for the project.}

\item{nameOfParameterIdentification}{A string representing the name of the parameter identification.}
}
\description{
This helper function updates fixed parameters in the project configuration by writing them to the appropriate sheet.
}
\keyword{internal}
