% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-config.R
\name{updateMappedPaths}
\alias{updateMappedPaths}
\title{Update Mapped Paths}
\usage{
updateMappedPaths(wb, linkedParameter, projectConfiguration)
}
\arguments{
\item{wb}{The workbook object to be updated.}

\item{linkedParameter}{A list of linked parameters.}

\item{projectConfiguration}{An object of class \code{ProjectConfigurationBMLM}, which contains configuration details for the project.}
}
\description{
This helper function updates the mapped paths in the workbook based on the linked parameters.
}
\keyword{internal}
