% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-startBMLMOptimization.R
\name{updateModelError}
\alias{updateModelError}
\title{Update Model Error}
\usage{
updateModelError(dtPrior, dtRes)
}
\arguments{
\item{dtPrior}{A data.table containing prior values.}

\item{dtRes}{A data.table containing results to update.}
}
\value{
A data.table containing updated results.
}
\description{
Updates the model error values in the results based on prior definitions.
}
\keyword{internal}
