% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-startBMLMOptimization.R
\name{updateOptimStatus}
\alias{updateOptimStatus}
\title{Update Optimization Status}
\usage{
updateOptimStatus(dtPrior, dtStartValues, optimEnv)
}
\arguments{
\item{dtPrior}{A data.table containing prior values.}

\item{dtStartValues}{A data.table containing start values.}

\item{optimEnv}{The optimization environment containing current state variables.}
}
\value{
A list containing the current optimization status.
}
\description{
Updates the optimization status with the current parameters and iteration.
}
\keyword{internal}
