% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-config.R
\name{updateOutputMappings}
\alias{updateOutputMappings}
\title{Update Output Mappings}
\usage{
updateOutputMappings(projectConfiguration, snp, selectedPI, wb)
}
\arguments{
\item{projectConfiguration}{An object of class \code{ProjectConfigurationBMLM}, which contains configuration details for the project.}

\item{snp}{A list representing the snapshot data.}

\item{selectedPI}{An integer representing the selected parameter identification index.}

\item{wb}{The workbook object to be updated.}
}
\description{
This helper function updates the output mappings in the workbook based on the project configuration and snapshot.
}
\keyword{internal}
