% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-startBMLMOptimization.R
\name{updateParameterValues}
\alias{updateParameterValues}
\title{Update Parameter Values}
\usage{
updateParameterValues(
  scenarioName,
  scenario,
  dtPrior,
  dtStartValues,
  dtMappedPaths
)
}
\arguments{
\item{scenarioName}{A string representing the name of the scenario.}

\item{scenario}{An object representing the scenario.}

\item{dtPrior}{A data.table containing prior values.}

\item{dtStartValues}{A data.table containing start values.}

\item{dtMappedPaths}{A data.table containing mapped paths for parameters.}
}
\value{
NULL (invisible).
}
\description{
This function updates parameter values in the scenario based on mapped paths, start values, and prior definitions.
}
\keyword{internal}
