% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-config.R
\name{updatePriorParameters}
\alias{updatePriorParameters}
\title{Update Prior Parameters}
\usage{
updatePriorParameters(dtPrior, dtDefinition, dataObserved, wb)
}
\arguments{
\item{dtPrior}{A data.table containing the existing prior parameters.}

\item{dtDefinition}{A data.table containing parameter definitions.}

\item{dataObserved}{A data.table of observed data.}

\item{wb}{An open workbook object where the updated prior parameters will be written.}
}
\value{
A data.table containing the updated prior parameters.
}
\description{
This function updates the prior parameters by adding global and hyper prior
parameters and model error parameters, then writes the updated data back
to the workbook.
}
\keyword{internal}
