% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-BMLMParameter.R
\name{validateAndLoadDefinition}
\alias{validateAndLoadDefinition}
\title{Validate and Load Definition}
\usage{
validateAndLoadDefinition(projectConfiguration)
}
\arguments{
\item{projectConfiguration}{An object of class ProjectConfiguration containing project configuration details.}
}
\value{
A data.table containing validated parameter definitions.
}
\description{
This function validates and loads parameter definitions from the BMLM configuration file.
}
\keyword{internal}
