% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-dataPreparation.R
\name{validateAndLoadPriorDefinition}
\alias{validateAndLoadPriorDefinition}
\title{Get Prior Definition}
\usage{
validateAndLoadPriorDefinition(projectConfiguration)
}
\arguments{
\item{projectConfiguration}{An object of class ProjectConfiguration containing project configuration details.}
}
\value{
A data.table with prior definitions.
}
\description{
This function retrieves prior definitions from the configuration file.
}
\keyword{internal}
