% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-enum.R
\docType{data}
\name{WEIGHTINGMODUS}
\alias{WEIGHTINGMODUS}
\title{WEIGHTINGMODUS for dataweights}
\format{
An object of class \code{list} of length 5.
}
\usage{
WEIGHTINGMODUS
}
\description{
WEIGHTINGMODUS for dataweights
}
\seealso{
Other enum helpers: 
\code{\link{ERRORMODEL}},
\code{\link{PARAMETERTYPE}}
}
\concept{enum helpers}
\keyword{datasets}
