% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-config.R
\name{addBMLMPConfiguration}
\alias{addBMLMPConfiguration}
\title{Create BMLM Project Configuration}
\usage{
addBMLMPConfiguration(
  projectConfiguration,
  nameOfParameterIdentfication = NULL,
  snapshotFile = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{projectConfiguration}{An object of class \code{ProjectConfiguration}, which contains configuration details for the project.
This object should include paths to necessary folders and files for the BMLM project.}

\item{nameOfParameterIdentfication}{A string representing the name of the parameter identification to be read from the snapshot.
This should match one of the names specified in the snapshot file. If NULL, the function will not attempt to read any parameters.}

\item{snapshotFile}{A string representing the path to the snapshot file. This file contains parameter identifications
that will be read and potentially used to update the project configuration. If NULL, no updates from a snapshot will occur.}

\item{overwrite}{A logical indicating whether to overwrite existing data in the configuration file if it already exists.
Default is FALSE, meaning existing data will not be overwritten unless specified.}
}
\value{
An object of class \code{ProjectConfiguration}, which includes the updated configuration details after processing.
}
\description{
This function creates and initializes a project configuration for BMLM (Biologically-Based Modeling and Learning Model).
It sets up the necessary configuration table and updates parameters based on a provided snapshot file.
}
