% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-checkProgress.R
\name{addLabel}
\alias{addLabel}
\title{Add Labels to Plot Data}
\usage{
addLabel(plotData, dtPrior, unitSep = " ")
}
\arguments{
\item{plotData}{A data.table containing the initial plot data.}

\item{dtPrior}{A data.table containing prior values for merging.}

\item{unitSep}{A string to separate the name and unit in the label. Default is a space.}
}
\value{
A data.table with labels added for each parameter.
}
\description{
This function merges the plot data with prior data to add labels for better visualization.
}
\keyword{internal}
