% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMLMOptimization.R
\docType{class}
\name{BMLMOptimization}
\alias{BMLMOptimization}
\title{BMLMOptimization}
\description{
An object storing configuration used project-wide
This class encapsulates the functionality for performing BMLM optimization.
It manages the configuration, output directories, logging, and optimization processes.
}
\section{Super class}{
\code{\link[ospsuite.utils:Printable]{ospsuite.utils::Printable}} -> \code{BMLMOptimization}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{runName}}{Identifier of optimization run.
Initialize BMLMOptimization}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BMLMOptimization-new}{\code{BMLMOptimization$new()}}
\item \href{#method-BMLMOptimization-startOptimization}{\code{BMLMOptimization$startOptimization()}}
\item \href{#method-BMLMOptimization-cleanUpStatus}{\code{BMLMOptimization$cleanUpStatus()}}
\item \href{#method-BMLMOptimization-print}{\code{BMLMOptimization$print()}}
\item \href{#method-BMLMOptimization-checkConvergence}{\code{BMLMOptimization$checkConvergence()}}
\item \href{#method-BMLMOptimization-checkParameterLimits}{\code{BMLMOptimization$checkParameterLimits()}}
\item \href{#method-BMLMOptimization-checkDistributions}{\code{BMLMOptimization$checkDistributions()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BMLMOptimization-new"></a>}}
\if{latex}{\out{\hypertarget{method-BMLMOptimization-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BMLMOptimization$new(
  projectConfiguration,
  runName,
  scenarioList,
  dataObserved,
  seed = 1234
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{projectConfiguration}}{A list containing project configuration details.}

\item{\code{runName}}{A character string representing the name of the current run.}

\item{\code{scenarioList}}{A list of scenarios to be optimized.}

\item{\code{dataObserved}}{A data.table containing the observed data to be fitted.}

\item{\code{seed}}{numeric value set random seed for startValue generation
Start the optimization process

This method initiates the optimization process using the specified method and control parameters.
Internally function \code{optim} is used, so please check help for more details}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BMLMOptimization-startOptimization"></a>}}
\if{latex}{\out{\hypertarget{method-BMLMOptimization-startOptimization}{}}}
\subsection{Method \code{startOptimization()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BMLMOptimization$startOptimization(
  projectConfiguration,
  method = "BFGS",
  control = list(),
  simulationRunOptions = NULL,
  failValue = 1e+10,
  lastStatusSavingIntervalInSecs = 60,
  startInBackground = TRUE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{projectConfiguration}}{A list containing project configuration details.}

\item{\code{method}}{A character string specifying the optimization method to be used (default is "BFGS").}

\item{\code{control}}{A list of control parameters for the optimization process.}

\item{\code{simulationRunOptions}}{Optional additional simulation run options.}

\item{\code{failValue}}{A numeric value to set if evaluation of the objective function fails (default is 1e+10).}

\item{\code{lastStatusSavingIntervalInSecs}}{An integer specifying the interval for saving the last status (default is 60 seconds).}

\item{\code{startInBackground}}{A logical indicating whether to start the optimization in the background (default is TRUE).}

\item{\code{...}}{Additional arguments to be passed to the optimization function \code{optim}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An invisible reference to the BMLMOptimization object.
Clean up the status of the optimization run

This method checks the status of the optimization run and updates it accordingly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BMLMOptimization-cleanUpStatus"></a>}}
\if{latex}{\out{\hypertarget{method-BMLMOptimization-cleanUpStatus}{}}}
\subsection{Method \code{cleanUpStatus()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BMLMOptimization$cleanUpStatus()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An invisible reference to the BMLMOptimization object.
Print properties and status of the BMLMOptimization object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BMLMOptimization-print"></a>}}
\if{latex}{\out{\hypertarget{method-BMLMOptimization-print}{}}}
\subsection{Method \code{print()}}{
This method prints the current properties and status of the BMLMOptimization object to the console.
Check Convergence of Model Parameters

This function checks the convergence of model parameters by reading a CSV file
containing convergence data and visualizing the results using ggplot2.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BMLMOptimization$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BMLMOptimization-checkConvergence"></a>}}
\if{latex}{\out{\hypertarget{method-BMLMOptimization-checkConvergence}{}}}
\subsection{Method \code{checkConvergence()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BMLMOptimization$checkConvergence(
  nPoints = 200,
  selectionMode = c("last", "random", "first")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{nPoints}}{An integer specifying the number of points to select for plotting.
Default is 200.}

\item{\code{selectionMode}}{A character string indicating the mode of selection for points.
Options are 'last'  'random', and 'first' (the first nPoints). Default is 'last'.
This function creates ggplot objects to display the current best and start values of the fitted parameter.
all values a display as relative between min and max value using defined scaling}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BMLMOptimization-checkParameterLimits"></a>}}
\if{latex}{\out{\hypertarget{method-BMLMOptimization-checkParameterLimits}{}}}
\subsection{Method \code{checkParameterLimits()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BMLMOptimization$checkParameterLimits(
  nCols = 2,
  nRows = 3,
  useInteractivePlots = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{nCols}}{An integer specifying the number of columns for the plot layout.}

\item{\code{nRows}}{An integer specifying the number of rows for the plot layout.}

\item{\code{useInteractivePlots}}{A Boolean, if TRUE plots will be generated with mouse overlay for absolute values.}

\item{\code{...}}{arguments passed on to function plotParameterLimits
This function creates ggplot objects to display the individual Values vs the fitted distributions.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BMLMOptimization-checkDistributions"></a>}}
\if{latex}{\out{\hypertarget{method-BMLMOptimization-checkDistributions}{}}}
\subsection{Method \code{checkDistributions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BMLMOptimization$checkDistributions(
  nCols = 2,
  nRows = 3,
  xScale = unlist(SCALING),
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{nCols}}{An integer specifying the number of columns for the plot layout.}

\item{\code{nRows}}{An integer specifying the number of rows for the plot layout.}

\item{\code{xScale}}{character 'Linear' or 'Log', scale of x axis}

\item{\code{...}}{additionla arguments passed on to plotDistributions
Logs a message to the specified log file with a timestamp.}
}
\if{html}{\out{</div>}}
}
}
}
