% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-BMLMParameter.R
\name{calculateLogLikelihood}
\alias{calculateLogLikelihood}
\title{Calculate Log Likelihood}
\usage{
calculateLogLikelihood(y, pred, model, sigma, isCensored, lloq, lowerBound)
}
\arguments{
\item{y}{A numeric vector of observed values.}

\item{pred}{A numeric vector of predicted values.}

\item{model}{A string indicating the error model type.}

\item{sigma}{A numeric value representing the standard deviation.}

\item{isCensored}{A logical indicating if the data is censored.}

\item{lloq}{A numeric value representing the lower limit of quantification.}

\item{lowerBound}{A numeric value representing the lower bound.}
}
\value{
A numeric value representing the log likelihood.
}
\description{
This function calculates the log likelihood based on observed and predicted values.
}
\keyword{internal}
