% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-config.R
\name{configurePriors}
\alias{configurePriors}
\title{Configure Priors}
\usage{
configurePriors(projectConfiguration, dataObserved, overwrite = FALSE)
}
\arguments{
\item{projectConfiguration}{An object of class \code{ProjectConfigurationBMLM}.}

\item{dataObserved}{A data.table of observed data.}

\item{overwrite}{A logical indicating whether to overwrite existing data. Default is FALSE.}
}
\value{
NULL (invisible).
}
\description{
This function configures priors based on the project configuration and observed data.
}
