% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-config.R
\name{createBMLMProjectConfiguration}
\alias{createBMLMProjectConfiguration}
\title{Title}
\usage{
createBMLMProjectConfiguration(
  projectPath = ".",
  snaphsotFile = NULL,
  PIName = NULL,
  modelFolder = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{overwrite}{}
}
\description{
Title
}
