% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotParameter.R
\name{createLineData}
\alias{createLineData}
\title{Create Line Data for Hyperparameters}
\usage{
createLineData(hyperParameter, xScale, dtValues = NULL)
}
\arguments{
\item{hyperParameter}{A data.table containing hyperparameter information.}

\item{xScale}{A character string specifying the scale of the x-axis ('linear' or 'log').}

\item{dtValues}{A data.table containing values for plotting.}
}
\value{
A data.table containing the line data for hyperparameters.
}
\description{
This function generates line data for hyperparameters based on their distributions.
}
\keyword{internal}
