% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-checkProgress.R
\name{createLineData}
\alias{createLineData}
\title{#' Create Line Data for Hyperparameters}
\usage{
createLineData(hyperParameter, dtPrior, xScale)
}
\arguments{
\item{hyperParameter}{A data.table containing hyperparameter information.}

\item{dtPrior}{A data.table containing prior values.}

\item{xScale}{A character string specifying the scale of the x-axis ('linear' or 'log').}
}
\value{
A data.table containing the line data for hyperparameters.
}
\description{
This function generates line data for hyperparameters based on their distributions.
}
\keyword{internal}
