% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-checkProgress.R
\name{createLineData}
\alias{createLineData}
\title{Create Line Data for Hyperparameters}
\usage{
createLineData(plotData, dtValues, xScale)
}
\arguments{
\item{plotData}{A data.table containing the data to be processed.}

\item{dtValues}{A data.table with ECDF calculated.}

\item{xScale}{A character string specifying the scale of the x-axis.}
}
\value{
A data.table containing the line data for hyperparameters.
}
\description{
This function generates the line data for hyperparameters based on their distributions.
}
\keyword{internal}
