% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotParameter.R
\name{customizeLegend}
\alias{customizeLegend}
\title{Customize Legend for ggplot Objects}
\usage{
customizeLegend(
  plotObject,
  colorScalingVector,
  aesthetics = c("color", "fill", "shape", "linetype")
)
}
\arguments{
\item{plotObject}{A ggplot object to which the legend will be added.}

\item{colorScalingVector}{A named vector of colors for different statuses.}
}
\value{
A ggplot object with a customized legend.
}
\description{
This function customizes the legend for ggplot objects by setting specific colors and shapes for different statuses.
}
\keyword{internal}
