% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-checkProgress.R
\name{customizeLegend}
\alias{customizeLegend}
\title{Customize Legend}
\usage{
customizeLegend(plotObject, colorScalingVector)
}
\arguments{
\item{plotObject}{A ggplot object to which the legend will be added.}

\item{colorScalingVector}{A named vector of colors for different statuses.}
}
\value{
A ggplot object with a customized legend.
}
\description{
This function customizes the legend for ggplot objects.
}
\keyword{internal}
