% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-enum.R
\docType{data}
\name{ERRORMODEL}
\alias{ERRORMODEL}
\title{Error Model}
\format{
An object of class \code{list} of length 3.
}
\usage{
ERRORMODEL
}
\description{
Error Model
}
\seealso{
Other enum helpers: 
\code{\link{PARAMETERTYPE}}
}
\concept{enum helpers}
\keyword{datasets}
