% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-config.R
\name{extractIdentificationParameter}
\alias{extractIdentificationParameter}
\title{Extract Identification Parameter}
\usage{
extractIdentificationParameter(linkedParameter)
}
\arguments{
\item{linkedParameter}{A list of linked parameters.}
}
\value{
A data.table containing the extracted parameters, including columns for parameter names, values, and their configurations.
}
\description{
This function extracts identification parameters from linked parameters.
}
\keyword{internal}
