% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-exportResult.R
\name{extractParameterValues}
\alias{extractParameterValues}
\title{Extract Parameter Values}
\usage{
extractParameterValues(dtNew, dtList, scenarioList)
}
\arguments{
\item{dtNew}{A data.table containing new parameter values to be extracted.}

\item{dtList}{A list of data.tables containing mapped paths and other relevant data.}

\item{scenarioList}{A list of scenarios, each containing simulation parameters.}
}
\value{
A data.table containing extracted parameter values along with their container paths and names.
}
\description{
This function extracts the optimized  parameter values and merges them with mapped paths.
}
\keyword{internal}
