% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-BMLMParameter.R
\name{getAllDistributions}
\alias{getAllDistributions}
\title{Get All Distributions}
\usage{
getAllDistributions()
}
\value{
A vector of unique distribution names.
}
\description{
This function retrieves all distributions available for prior definitions and hyperprameter.
}
