% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-scenarioSimulation.R
\name{getDefaultPopulationLine}
\alias{getDefaultPopulationLine}
\title{Title}
\usage{
getDefaultPopulationLine(projectConfig, params, modelFile)
}
\arguments{
\item{modelFile}{}
}
\description{
Title
}
