% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-BMLMParameter.R
\name{getDistributionParameters}
\alias{getDistributionParameters}
\title{Get Distribution Parameters}
\usage{
getDistributionParameters(distributionName)
}
\arguments{
\item{distributionName}{A string representing the name of the distribution.}
}
\value{
A vector of parameters for the specified distribution.
}
\description{
This function retrieves the parameters for a specified distribution.
}
