% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-BMLMParameter.R
\name{getDistributionParameters}
\alias{getDistributionParameters}
\title{Title}
\usage{
getDistributionParameters(distributionName)
}
\arguments{
\item{distributionName}{}
}
\description{
Title
}
