% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-dataPreparation.R
\name{getHyperParameter}
\alias{getHyperParameter}
\title{Get Hyper Parameters}
\usage{
getHyperParameter(dtPrior)
}
\arguments{
\item{dtPrior}{A data.table containing prior definitions.}
}
\value{
A data.table containing hyperparameters and their distributions.
}
\description{
This function retrieves hyperparameters from the prior definition.
}
\keyword{internal}
