% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-evaluateResult.R
\name{loadListsForRun}
\alias{loadListsForRun}
\title{Load Lists for Run}
\usage{
loadListsForRun(projectConfiguration, runName)
}
\arguments{
\item{projectConfiguration}{An object of class projectConfiguration containing configuration settings, including
the output folder path where the CSV files are located.}

\item{runName}{A character string representing the name of the run, which is used to
construct the output directory path.}
}
\value{
A list of data.tables, each corresponding to one of the CSV files loaded. The
list elements are named after the CSV files (without the '.csv' extension).
}
\description{
This function loads various CSV files from a specified output directory for a given name of a BLMLM run,
processes them into data.tables, and returns them as a list.
}
