% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-distribution.R
\name{lognormal_geomean}
\alias{lognormal_geomean}
\alias{dlnorm_geomean}
\alias{plnorm_geomean}
\alias{qlnorm_geomean}
\alias{rlnorm_geomean}
\title{Log-Normal Functions with Geometric Mean}
\usage{
dlnorm_geomean(x, geomean = 1, geosd = 1, log = FALSE)
}
\arguments{
\item{x}{A numeric vector of quantiles (for \code{dlnorm_geomean}).}

\item{geomean}{A numeric value representing the geometric mean. Default is 1.}

\item{geosd}{A numeric value representing the standard deviation of the logarithm. Default is 1.}

\item{log}{A logical value indicating whether to return the logarithm of the density. Default is FALSE.}

\item{q}{A numeric vector of quantiles (for \code{plnorm_geomean}).}

\item{p}{A numeric vector of probabilities (for \code{qlnorm_geomean}).}

\item{n}{A numeric value indicating the number of random values to generate (for \code{rlnorm_geomean}).}

\item{lower.tail}{A logical value indicating whether to return the lower tail probability. Default is TRUE.}

\item{log.p}{A logical value indicating whether to return the logarithm of the probability. Default is FALSE.}
}
\value{
A numeric value or vector depending on the function:
\describe{
\item{dlnorm_geomean}{Density at the specified point.}
\item{plnorm_geomean}{Cumulative probability at the specified quantiles.}
\item{qlnorm_geomean}{Quantiles corresponding to the specified probabilities.}
\item{rlnorm_geomean}{Random values generated from the log-normal distribution.}
}
}
\description{
These functions provide an interface to the log-normal distribution using
the geometric mean and geometric standard deviation instead of the log mean and log sd. They internally call the
corresponding base R functions: \code{dlnorm}, \code{plnorm}, \code{qlnorm}, and \code{rlnorm}.
e.g. dlnorm_geomean(geomean, geosd, x = x) calls dlnorm(logmean = log(geomean), logsd = log(geosd), x = x)
}
\section{Functions}{

\describe{
\item{\code{dlnorm_geomean(geomean, geosd, ...)}}{Density of the log-normal distribution.}
\item{\code{plnorm_geomean(q, geomean, geosd, ...)}}{Cumulative distribution function of the log-normal distribution.}
\item{\code{qlnorm_geomean(p, geomean, geosd, ...)}}{Quantile function of the log-normal distribution.}
\item{\code{rlnorm_geomean(n, geomean, geosd, ...)}}{Random generation from the log-normal distribution.}
}
}

