% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-startBMLMOptimization.R
\name{optimizeParameters}
\alias{optimizeParameters}
\title{Optimize Parameters}
\usage{
optimizeParameters(
  dtList,
  scenarioList,
  simulationRunOptions,
  method,
  control,
  outputDir,
  lastStatusSavingInterval,
  bestValue,
  logFile,
  startTime,
  failValue,
  ...
)
}
\arguments{
\item{dtList}{A list containing data tables for optimization.}

\item{scenarioList}{A list of scenarios to be optimized.}

\item{simulationRunOptions}{A list of options for simulation runs.}

\item{method}{A character string specifying the optimization method.}

\item{control}{A list of control parameters for the optimization function.}

\item{outputDir}{A character string representing the path to the output directory.}

\item{lastStatusSavingInterval}{An integer specifying the interval for saving the last status.}

\item{bestValue}{A numeric value representing the current best value of optimization.}

\item{logFile}{A character string representing the file path for the log file.}

\item{startTime}{A POSIXct object representing the start time of the optimization process.}

\item{failValue}{A numeric value to set if evaluation of the objective function fails.}

\item{...}{Additional arguments passed to the optimization function.}
}
\value{
The result of the optimization process.
}
\description{
Performs the optimization of parameters using the specified method and control settings.
}
