% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-checkProgress.R
\name{plotConvergence}
\alias{plotConvergence}
\title{Check Convergence of Model Parameters}
\usage{
plotConvergence(
  dtConvergence,
  displayVariablesIndx = NULL,
  titletxt = NULL,
  nPoints = 200,
  selectionMode = c("last", "random", "first")
)
}
\arguments{
\item{dtConvergence}{A data.table containing the convergence data.}

\item{displayVariablesIndx}{An integer vector specifying which variables to display.}

\item{titletxt}{A string to include in the plot title.}

\item{nPoints}{An integer specifying the number of points to select for plotting. Default is 200.}

\item{selectionMode}{A character string indicating the mode of selection for points.
Options are 'last', 'random', and 'first'. Default is 'last'.}
}
\value{
A ggplot object visualizing the convergence of model parameters.
}
\description{
This function checks the convergence of model parameters by reading a data.table
containing convergence data and visualizing the results using ggplot2.
}
