% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-checkProgress.R
\name{plotConvergence}
\alias{plotConvergence}
\title{Check Convergence of Model Parameters}
\usage{
plotConvergence(
  dtConvergence,
  titletxt = NULL,
  nPoints = 200,
  selectionMode = c("last", "random", "first")
)
}
\arguments{
\item{dtConvergence}{A data.table containing the loglikelihood vsplot titlename of the run.}

\item{nPoints}{An integer specifying the number of points to select for plotting.
Default is 200.}

\item{selectionMode}{A character string indicating the mode of selection for points.
Options are 'last', 'random', and 'first'. Default is 'last'.}
}
\value{
A ggplot object visualizing the convergence of model parameters.
The plot shows the values of logTimeProfile, logHyperParameter, logPrior,
and objectiveValue over iterations.
}
\description{
This function checks the convergence of model parameters by reading a CSV file
containing convergence data and visualizing the results using ggplot2.
}
\keyword{internal}
