% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCorrelation.R
\name{plotCorrelations}
\alias{plotCorrelations}
\title{Plot Correlations}
\usage{
plotCorrelations(
  dtList,
  statusList,
  titeltxt = NULL,
  method = "spearman",
  statusToShow = c("best", "current", "start"),
  scenarioList,
  corCut = 0.5,
  pValueCut = 0.1
)
}
\arguments{
\item{dtList}{A list containing prior and start values data tables.}

\item{statusList}{A list containing the current and best parameter values.}

\item{titeltxt}{A string to include in the plot title.}

\item{method}{A character string specifying the correlation method to use (default is 'spearman').
Input of function \code{stats::cor()}}

\item{statusToShow}{A character string indicating which status to show. Options are 'best', 'current', and 'start'.}

\item{scenarioList}{A list of scenarios to analyze.}

\item{corCut}{A numeric value for the correlation cutoff threshold. Default is 0.5.}

\item{pValueCut}{A numeric value for the Chi-squared cutoff threshold. Default is 0.1.}
}
\value{
NULL This function does not return a value; it prints the correlation plot directly.
}
\description{
This function generates a correlation matrix plot for the specified status and checks
for relevant columns in the population data.
}
\seealso{
Other plotting: 
\code{\link{defaultColorsParameterPlots}()},
\code{\link{plotBestVsStartParameter}()},
\code{\link{plotConvergence}()},
\code{\link{plotDistributions}()},
\code{\link{plotParameterLimits}()},
\code{\link{plotParameterValuesVsPrior}()},
\code{\link{plotPredictedVsObserved}()},
\code{\link{plotPredictedVsTime}()},
\code{\link{plotResidualsAsHistogram}()},
\code{\link{plotResidualsAsQQ}()},
\code{\link{plotResidualsDistribution}()},
\code{\link{plotResidualsVsTime}()}
}
\concept{plotting}
