% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotParameter.R
\name{plotDistributions}
\alias{plotDistributions}
\title{Plot Distributions}
\usage{
plotDistributions(
  dtList,
  statusList,
  nCols = 2,
  nRows = 3,
  xScale = unlist(SCALING),
  titeltxt = NULL,
  parameterFilter = NULL,
  zoomOnData = FALSE,
  colorScalingVector = defaultColorsParameterPlots()
)
}
\arguments{
\item{dtList}{A data.table containing the data to be plotted.}

\item{statusList}{A list containing the current and best parameter values.}

\item{nCols}{An integer specifying the number of columns for faceting. Default is 2.}

\item{nRows}{An integer specifying the maximum number of rows for faceting. Default is 3.}

\item{xScale}{A character string specifying the scale of the x-axis. Default is 'log'.}

\item{titeltxt}{A string to include in the plot title.}

\item{parameterFilter}{A character vector for filtering parameter names. Default is NULL.}

\item{zoomOnData}{A logical value indicating whether to zoom in on data. Default is FALSE.}

\item{colorScalingVector}{A named vector of colors for different statuses.
Default is set by \code{defaultColorsParameterPlots()}.}
}
\value{
Returns an invisible NULL after printing the plots.
}
\description{
This function creates a series of plots to visualize the cumulative distribution of individual values and hyperparameters.
}
\seealso{
Other plotting: 
\code{\link{defaultColorsParameterPlots}()},
\code{\link{plotBestVsStartParameter}()},
\code{\link{plotConvergence}()},
\code{\link{plotCorrelations}()},
\code{\link{plotParameterLimits}()},
\code{\link{plotParameterValuesVsPrior}()},
\code{\link{plotPredictedVsObserved}()},
\code{\link{plotPredictedVsTime}()},
\code{\link{plotResidualsAsHistogram}()},
\code{\link{plotResidualsAsQQ}()},
\code{\link{plotResidualsDistribution}()},
\code{\link{plotResidualsVsTime}()}
}
\concept{plotting}
