% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-checkProgress.R
\name{plotDistributions}
\alias{plotDistributions}
\title{Plot Distributions}
\usage{
plotDistributions(
  plotData,
  nCols = 2,
  nRows = 3,
  xScale = unlist(SCALING),
  colorScalingVector = c(start = "lightblue", current = "darkgreen", best = "orange")
)
}
\arguments{
\item{plotData}{A data.table containing the data to be plotted.}

\item{nCols}{An integer specifying the number of columns for faceting. Default is 2.}

\item{nRows}{An integer specifying the maximum number of rows for faceting. Default is 3.}

\item{xScale}{A character string specifying the scale of the x-axis. Default is 'Log'}

\item{colorScalingVector}{A named vector of colors for different statuses. Default includes light blue, dark green, and orange.}
}
\value{
Returns an invisible NULL after printing the plots.
}
\description{
This function creates a series of plots to visualize the cumulative distribution of individual values and hyperparameters.
}
