% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotParameter.R
\name{plotParameterLimits}
\alias{plotParameterLimits}
\title{Create and Print Parameter Limits Plots}
\usage{
plotParameterLimits(
  dtList,
  statusList,
  titeltxt,
  nCols,
  nRows,
  colorScalingVector = defaultColorsParameterPlots()
)
}
\arguments{
\item{dtList}{A list containing prior and start values data tables.}

\item{statusList}{A list containing the current and best parameter values.}

\item{titeltxt}{A string to include in the plot title.}

\item{nCols}{An integer specifying the number of columns for the plot layout.}

\item{nRows}{An integer specifying the number of rows for the plot layout.}

\item{colorScalingVector}{A named vector of colors for different statuses.
Default is set by \code{defaultColorsParameterPlots()}.}
}
\value{
invisible list of ggplot objects.
}
\description{
This function creates ggplot objects to display the current best and start values
of the fitted parameters.
}
