% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-checkProgress.R
\name{plotParameterLimits}
\alias{plotParameterLimits}
\title{Create and Print Parameter Limits Plots}
\usage{
plotParameterLimits(
  dtList,
  statusList,
  titeltxt,
  nCols,
  nRows,
  colorScalingVector = c(current = "darkgreen", start = "lightblue", best = "orange")
)
}
\arguments{
\item{dtList}{A list containing prior and start values data tables.}

\item{statusList}{A list containing the current and best parameter values.}

\item{titeltxt}{A string to include in the plot title.}

\item{nCols}{An integer specifying the number of columns for the plot layout.}

\item{nRows}{An integer specifying the number of rows for the plot layout.}

\item{colorScalingVector}{A named vector of colors for different statuses.
Default includes dark green, light blue, and orange.}
}
\value{
NULL This function does not return a value; it prints the plots directly.
}
\description{
This function creates ggplot objects to display the current best and start values
of the fitted parameters.
}
