% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-checkProgress.R
\name{plotParameterLimits}
\alias{plotParameterLimits}
\title{Create Plots}
\usage{
plotParameterLimits(
  plotData,
  titeltxt,
  nCols,
  nRows,
  useInteractivePlots = FALSE,
  colorScalingVector = c(start = "lightblue", current = "darkgreen", best = "orange")
)
}
\arguments{
\item{plotData}{A data.table containing the reshaped plot data.}

\item{titeltxt}{A string  to include in the plot title.}

\item{nCols}{An integer specifying the number of columns for the plot layout.}

\item{nRows}{An integer specifying the number of rows for the plot layout.}
}
\value{
A list of ggplot objects representing the plots for the parameter values.
}
\description{
This function creates ggplot objects based on the prepared plot data.
If the number of plots exceeds the specified grid size, it splits the plots into multiple sets.
}
\keyword{internal}
