% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPredicted.R
\name{plotPredictedVsObserved}
\alias{plotPredictedVsObserved}
\title{Create and Print Predicted vs Observed Plots}
\usage{
plotPredictedVsObserved(
  filteredData,
  addRegression = TRUE,
  xyScale = unlist(SCALING),
  ...
)
}
\arguments{
\item{filteredData}{A data frame containing the data to be plotted. It should include columns for
scenario, outputPathId, group, predicted, yValues, and isCensored.}

\item{addRegression}{A logical value indicating whether to add regression lines to the plot (default is TRUE).}

\item{xyScale}{A character string specifying the scale type for the x and y axes (default is "log").}

\item{...}{Additional arguments passed on to ospsuite.plots::plotPredVsObs.}
}
\value{
An invisible ggplot object for the predicted vs observed plot.
}
\description{
This function generates a plot to compare predicted values against time for a single outputPathId and scenario.
}
\seealso{
Other plotting: 
\code{\link{defaultColorsParameterPlots}()},
\code{\link{plotBestVsStartParameter}()},
\code{\link{plotConvergence}()},
\code{\link{plotCorrelations}()},
\code{\link{plotDistributions}()},
\code{\link{plotParameterLimits}()},
\code{\link{plotParameterValuesVsPrior}()},
\code{\link{plotPredictedVsTime}()},
\code{\link{plotResidualsAsHistogram}()},
\code{\link{plotResidualsAsQQ}()},
\code{\link{plotResidualsDistribution}()},
\code{\link{plotResidualsVsTime}()}
}
\concept{plotting}
