% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-checkProgress.R
\name{plotPredictedVsObserved}
\alias{plotPredictedVsObserved}
\title{Create and Print Predicted vs Observed Plots}
\usage{
plotPredictedVsObserved(
  dtRes,
  addRegression = TRUE,
  xyScale = unlist(SCALING),
  nCols = 2,
  titeltxt = NULL,
  ...
)
}
\arguments{
\item{dtRes}{A data frame containing the data to be plotted. It should include columns for
scenario, outputPathId, and group.}

\item{addRegression}{A logical value indicating whether to add regression lines to the plot.}

\item{xyScale}{A character string specifying the scale type for the x and y axes (default "log").}

\item{nCols}{An integer specifying the number of columns for the facet wrap. (Default = 2)}

\item{titeltxt}{A string to include in the plot title.}

\item{...}{Additional arguments passed on to ospsuite.plots::plotPredVsObs}
}
\value{
NULL This function does not return a value; it prints the plots directly.
}
\description{
This function generates a plot for each unique outputPathId in the provided dataset.
Each plot displays predicted vs observed values and includes facets for scenario and group.
}
