% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-checkProgress.R
\name{plotResidualsAsQQ}
\alias{plotResidualsAsQQ}
\title{Create and Print Residuals as QQ Plot}
\usage{
plotResidualsAsQQ(dtRes, nCols = 2, titeltxt = NULL, self...)
}
\arguments{
\item{dtRes}{A data frame containing the data to be plotted. It should include columns for
scenario, outputPathId, group, and the residuals.}

\item{nCols}{An integer specifying the number of columns for the facet wrap. (Default = 2)}

\item{titeltxt}{A string to include in the plot title.}

\item{...}{Additional arguments passed on to \code{ospsuite.plots::plotQQ}.}
}
\value{
NULL This function does not return a value; it prints the plots directly.
}
\description{
This function generates a QQ plot to assess the normality of the residuals.
}
