% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPredicted.R
\name{plotResidualsDistribution}
\alias{plotResidualsDistribution}
\title{Plot Residual Distribution}
\usage{
plotResidualsDistribution(filteredData, ...)
}
\arguments{
\item{filteredData}{A data.table containing the residuals and other relevant data.}

\item{...}{Additional arguments passed to the plotting function.}
}
\value{
A ggplot object visualizing the residual distribution.
}
\description{
This function creates a plot to visualize the distribution of residuals from the model predictions.
}
