% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPredicted.R
\name{plotResidualsVsTime}
\alias{plotResidualsVsTime}
\title{Create and Print Residuals vs Time Plots}
\usage{
plotResidualsVsTime(filteredData, ...)
}
\arguments{
\item{filteredData}{A data frame containing the data to be plotted. It should include columns for
scenario, outputPathId, group, and the residuals.}

\item{...}{Additional arguments passed on to ospsuite.plots::plotResVsCov.}
}
\value{
A ggplot object for the residuals vs time plot.
}
\description{
This function generates a plot for residuals versus observed values for a single outputPathId.
}
