% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-checkProgress.R
\name{plotResidualsVsTime}
\alias{plotResidualsVsTime}
\title{Create and Print Residuals vs Time Plots}
\usage{
plotResidualsVsTime(dtRes, nCols = 2, titeltxt = NULL, ...)
}
\arguments{
\item{dtRes}{A data frame containing the data to be plotted. It should include columns for
scenario, outputPathId, group, and the residuals.}

\item{nCols}{An integer specifying the number of columns for the facet wrap. (Default = 2)}

\item{titeltxt}{A string to include in the plot title.}

\item{...}{Additional arguments passed on to ospsuite.plots::plotResVsCov.}
}
\value{
NULL This function does not return a value; it prints the plots directly.
}
\description{
This function generates a plot for residuals versus observed values from the provided dataset.
Each plot includes facets for scenario and group, and is created for each unique outputPathId.
}
