% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-BMLMParameter.R
\name{prepareInputData}
\alias{prepareInputData}
\title{Prepare Input Data for L-BFGS-B Algorithm}
\usage{
prepareInputData(dtPrior, dtStartValues, dtDefinition)
}
\arguments{
\item{dtPrior}{A data.table containing prior values with columns: 'value', 'MinValue', 'MaxValue', 'Scaling', 'valueMode'.}

\item{dtStartValues}{A data.table containing start values with at least a 'Name' column.}

\item{dtDefinition}{A data.table containing definitions with columns: 'Name', 'Scaling'.}
}
\value{
A data.table with combined and transformed input data for the L-BFGS-B algorithm.
}
\description{
This function prepares the input data for the L-BFGS-B algorithm by merging and transforming
the provided data tables. It handles log transformations based on specified conditions.
}
\keyword{internal}
