% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-checkProgress.R
\name{preparePlotDataParameterValues}
\alias{preparePlotDataParameterValues}
\title{Prepare Plot Data for Parameter Values}
\usage{
preparePlotDataParameterValues(dtList, currentStatus, bestStatus)
}
\arguments{
\item{dtList}{A list containing prior and start values data tables.}

\item{currentStatus}{A list which contains the current values.}

\item{bestStatus}{A list which contains the best values.}
}
\value{
A data.table containing the prepared plot data.
}
\description{
This function prepares the data for plotting by loading the necessary status files
and transforming the data into a suitable format for visualization.
}
\keyword{internal}
