% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-startBMLMOptimization.R
\name{processScenario}
\alias{processScenario}
\title{Process a Scenario}
\usage{
processScenario(scenarioName, scenario, dtList, simulationRunOptions)
}
\arguments{
\item{scenarioName}{A character string representing the name of the scenario to be processed.}

\item{scenario}{A list containing the parameters and settings specific to the scenario.}

\item{dtList}{A list containing data tables:
\itemize{
\item \code{prior}: Data table for prior parameters.
\item \code{startValues}: Data table for starting values.
\item \code{mappedPaths}: Data table for mapped paths.
\item \code{data}: Data table used for predictions.
}}

\item{simulationRunOptions}{A list containing options for running the simulation, which may include settings such as the number of iterations, random seed, etc.}
}
\value{
A list containing:
\item{result}{A list with the results of the scenario run.}
\item{predictions}{A data frame or list of predictions generated from the scenario results.}
}
\description{
This function updates parameter values for a given scenario, runs the scenario using the specified simulation options,
and retrieves predictions based on the scenario results.
}
