% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-IdentificationParameter.R,
%   R/utilities-config.R
\name{readIdentificationParameterFromSnapshot}
\alias{readIdentificationParameterFromSnapshot}
\title{Title}
\usage{
readIdentificationParameterFromSnapshot(
  snaphsotFile,
  PIName,
  projectConfig,
  overwrite = FALSE
)

readIdentificationParameterFromSnapshot(
  snaphsotFile,
  PIName,
  projectConfig,
  overwrite = FALSE
)
}
\arguments{
\item{PIName}{}
}
\description{
Title

Title
}
