% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-checkProgress.R
\name{reshapePlotDataParameterValues}
\alias{reshapePlotDataParameterValues}
\title{Reshape Plot Data}
\usage{
reshapePlotDataParameterValues(plotData)
}
\arguments{
\item{plotData}{A data.table containing the initial plot data.}

\item{dtList}{A list containing prior and start values data tables.}
}
\value{
A data.table with enhanced plot data.
}
\description{
This function reshapes the data.table and adds additional columns to the plot data for labeling and categorization.
}
\keyword{internal}
