% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-enum.R
\docType{data}
\name{SCALING}
\alias{SCALING}
\title{Scalings}
\format{
An object of class \code{list} of length 2.
}
\usage{
SCALING
}
\description{
Scalings
}
\seealso{
Other enum helpers: 
\code{\link{ERRORMODEL}},
\code{\link{PARAMETERTYPE}},
\code{\link{RUNSTATUS}}
}
\concept{enum helpers}
\keyword{datasets}
