% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-BMLMParameter.R
\name{setParameterToTables}
\alias{setParameterToTables}
\title{Set Parameter to Tables}
\usage{
setParameterToTables(dtList, params, logScaleIndex)
}
\arguments{
\item{dtList}{A list containing various data.tables used in the optimization process.}

\item{params}{A numeric vector of parameters for the likelihood calculation.}

\item{logScaleIndex}{A logical vector indicating which parameters are on a log scale.}
}
\value{
A list containing the updated data.tables.
}
\description{
This function updates the parameter values in the provided data.tables based on the optimization results.
}
\keyword{internal}
