% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-BMLMParameter.R
\name{startBMLMOptimization}
\alias{startBMLMOptimization}
\title{Start BMLM Optimization}
\usage{
startBMLMOptimization(
  projectConfiguration,
  scenarioList,
  dataObserved,
  seed = 1234,
  method = "L-BFGS-B",
  control = list(),
  simulationRunOptions = NULL
)
}
\arguments{
\item{projectConfiguration}{An object of class ProjectConfiguration containing project configuration details, including the path to the BMLM configuration file.}

\item{scenarioList}{A list of scenarios to be used in the optimization.}

\item{dataObserved}{A data.table containing observed data.}

\item{seed}{A numeric seed for random number generation (default is 1234).}

\item{method}{The optimization method to be used (default is "L-BFGS-B").}

\item{control}{A list of control parameters for the optimization.}

\item{simulationRunOptions}{an object of class ospsuite::SimulationRunOptions. Optional parameters for running simulations.}
}
\value{
An invisible NULL. The results of the optimization process are saved in files.
}
\description{
This function initiates the BMLM optimization process based on project configuration and observed data.
}
