% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-startBMLMOptimization.R
\name{startBMLMOptimization}
\alias{startBMLMOptimization}
\title{Start BMLM Optimization}
\usage{
startBMLMOptimization(
  projectConfiguration,
  runName,
  scenarioList,
  dataObserved,
  seed = 1234,
  method = "L-BFGS-B",
  control = list(),
  simulationRunOptions = NULL,
  restartAtBestValues = FALSE,
  failValue = -1e+10,
  lastStatusSavingInterval = 60,
  startInBackground = FALSE,
  ...
)
}
\arguments{
\item{projectConfiguration}{A list containing project configuration details.}

\item{runName}{A character string representing the name of the current run.}

\item{scenarioList}{A list of scenarios to be optimized.}

\item{dataObserved}{A data.table containing the observed data to be fitted.}

\item{seed}{An integer for random number generation (default is 1234).}

\item{method}{A character string specifying the optimization method (default is "L-BFGS-B").}

\item{control}{A list of control parameters for the optimization function passed to \code{optim}.}

\item{simulationRunOptions}{A list of options for simulation runs. Specify the parameters for the simulation here.}

\item{restartAtBestValues}{A logical indicating whether to restart from the best values found (default is FALSE).}

\item{failValue}{A numeric value to set if evaluation of the objective function fails (default is -1e+10).}

\item{lastStatusSavingInterval}{An integer specifying the interval for saving the last status (default is 60 seconds).}

\item{startInBackground}{A logical indicating whether to start the optimization in the background (default is FALSE).}

\item{...}{Additional arguments passed to the \code{optim}.}
}
\value{
Returns NULL invisibly after the optimization process is complete. This indicates that the function is primarily used for its side effects, such as saving results and logging.
}
\description{
This function initiates the BMLM optimization process based on the provided configuration and parameters.
The core of the optimization is performed using the \code{optim} function, so for more help on additional parameters
and control options, users should refer to the documentation for \code{optim}.
The optimized parameters and the final value of the objective function are saved to an RDS file, along with
relevant log information in a text file.
}
