% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-dataPreparation.R
\name{validateAndLoadIndividualStartValues}
\alias{validateAndLoadIndividualStartValues}
\title{Get Individual Start Values}
\usage{
validateAndLoadIndividualStartValues(
  projectConfiguration,
  dtHyperParameter,
  dataObserved,
  seed
)
}
\arguments{
\item{projectConfiguration}{An object of class ProjectConfiguration containing project configuration details.}

\item{dtHyperParameter}{A data.table containing hyperparameters and their distributions.}

\item{dataObserved}{A data.table containing observed data.}

\item{seed}{A numeric seed for random number generation.}
}
\value{
A data.table with updated start values.
}
\description{
This function reads individual start values from an Excel file and adjusts them based on hyperparameters.
}
\keyword{internal}
