% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-dataPreparation.R
\name{validateAndLoadMappedPaths}
\alias{validateAndLoadMappedPaths}
\title{Validate and Load Mapped Paths}
\usage{
validateAndLoadMappedPaths(
  projectConfiguration,
  dtPrior,
  dtStartValues,
  scenarioList
)
}
\arguments{
\item{projectConfiguration}{An object of class ProjectConfiguration containing project configuration details.}

\item{scenarioList}{A list of scenarios to be used in the optimization.}
}
\value{
A data.table containing mapped paths for parameters.
}
\description{
This function validates and loads mapped paths from the BMLM configuration file.
}
\keyword{internal}
