% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-BMLMParameter.R
\name{validateAndLoadPriorDefinition}
\alias{validateAndLoadPriorDefinition}
\title{Get Prior Definition}
\usage{
validateAndLoadPriorDefinition(projectConfiguration, dtDefinition)
}
\arguments{
\item{projectConfiguration}{An object of class ProjectConfiguration containing project configuration details.}

\item{dtDefinition}{A data.table containing parameter definitions.}
}
\value{
A data.table with prior definitions.
}
\description{
This function retrieves prior definitions from the configuration file.
}
\keyword{internal}
