% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-config.R
\name{validateParameterDefinition}
\alias{validateParameterDefinition}
\title{Validate Parameter Definition}
\usage{
validateParameterDefinition(dtDefinition)
}
\arguments{
\item{dtDefinition}{A data.table containing parameter definitions to be validated.}
}
\value{
invisible(NULL). This function is called for its side effects (validation).
}
\description{
This function validates the parameter definitions to ensure that the
required columns are present and contain valid data.
}
\keyword{internal}
